% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{add_mouse}
\alias{add_mouse}
\title{Add mouse object to an experiment}
\usage{
add_mouse(e, m, replace = FALSE)
}
\arguments{
\item{e}{experiment object}

\item{m}{mouse object}

\item{replace}{(bool, default = FALSE) Replace a mouse already contained in an experiment object.}
}
\value{
an experiment object
}
\description{
This function takes an experiment object and mouse object and adds the mouse object to the experiment. The mouse's unprocessed data, including all of it's individual
slice information and raw imported segmentation and registration data will be not be added from the mouse object to save space. Any desire to modify this data must be done at the mouse
object level before continuing further.

This function will also read the individual mouse attributes and automatically populate the experimental attributes that are relevant. For example, the 'group' attribute of
a mouse will be read and automatically added to the experiment object's 'experimental_groups' attribute if it is a new unique experimental group name.
}
\examples{
e <- experiment(experiment_name = "example")
m <- mouse(mouse_ID = "Mouse1")
e <- add_mouse(e, m)
e <- add_mouse(e, m, replace = TRUE)
}
