% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attr2match}
\alias{attr2match}
\title{attr2match}
\format{
A list
\describe{
}
}
\usage{
attr2match
}
\description{
A custom list to match attributes of a mouse and experiment object
respectively.
}
\keyword{datasets}
