% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_functions.R
\name{check_ontology_coding}
\alias{check_ontology_coding}
\title{Checks the acronyms and full length region names to match with internally stored ontology}
\usage{
check_ontology_coding(e, ontology = "allen")
}
\arguments{
\item{e}{experiment object}

\item{ontology}{(str, default = "allen") Region ontology to check against. Options = "allen" or "unified"}
}
\value{
e experiment object
}
\description{
Run this as a quality check after importing an external dataset uising \code{\link[=import_mapped_datasets]{import_mapped_datasets()}}. This goes through
all dataframes for all channels imported and replaces any non-matching acronyms and region names with those as they are exactly coded in SMARTR.
}
\examples{
e <- check_ontology_coding(e, "allen")
}
