% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_redundant_parents}
\alias{check_redundant_parents}
\title{Check for redundant parent regions included in a list of acronyms in a plate. For example, if all the the subregions for
the hypothalamus are represented, the HY should not be included in the list.}
\usage{
check_redundant_parents(acronyms, ontology = "allen")
}
\arguments{
\item{acronyms}{(vec) a vector of acronyms to check for possible parents that are redundantly included in the vector.}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}
}
\value{
A list containing two elements: one vector of unique child acronyms, a vector of the parent regions considered redundant
}
\description{
Check for redundant parent regions included in a list of acronyms in a plate. For example, if all the the subregions for
the hypothalamus are represented, the HY should not be included in the list.
}
\examples{
check_redundant_parents(acronyms = c("ACA5", "ACA1", "ACA"))
check_redundant_parents(acronyms = c("VMHDM", "VMHSh", "VMH"), ontology = "unified")
}
