% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{combine_norm_cell_counts}
\alias{combine_norm_cell_counts}
\title{Combine cell counts across the mice in an experiment}
\usage{
combine_norm_cell_counts(e, by = c("groups", "sex"))
}
\arguments{
\item{e}{experiment object}

\item{by}{(str) names of the experiment attributes used to divide up the cell counts. Will be same attributes to group by during analysis.}
}
\value{

}
\description{
Combine cell counts across the mice in an experiment
}
\examples{
e <- combine_norm_cell_counts(e, by = c('groups', 'sex'))
}
