% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{correlation_diff_permutation}
\alias{correlation_diff_permutation}
\title{This function performs a permutation analysis to compare the region pairwise correlation coefficients between two different analysis groups
specified by \code{correlation_list_name_1} and \code{correlation_list_name_2.} Note that both of these analysis groups must have the same
number of channels to compare. The functions \code{get_correlations()} needs to have been run for each of these analysis groups prior to
running this function.}
\usage{
correlation_diff_permutation(
  e,
  correlation_list_name_1 = "AD",
  correlation_list_name_2 = "control",
  n_shuffle = 1000,
  alpha = 0.05,
  p_adjust_method = "BH",
  seed = 5,
  channels = c("cfos", "eyfp", "colabel"),
  ...
)
}
\arguments{
\item{e}{experiment object}

\item{correlation_list_name_1}{(default = "AD") The name of the correlation list object used as the first group for comparison.}

\item{correlation_list_name_2}{(default = "control") The name of the correlation list object used as the second group for comparison.}

\item{n_shuffle}{(int, default = 1000) The number of permutation shuffles.}

\item{alpha}{(float, default = 0.05) The alpha cutoff for significance between region pairwise correlation differences}

\item{p_adjust_method}{(bool or str, default = "BH") Benjamini-Hochberg method is recommended.
Apply the named method to control for the inflated false discovery rate or FWER. Set to FALSE or "none"
to keep "raw" p values. See also \code{\link[stats:p.adjust]{stats::p.adjust()}} for the correction options.}

\item{seed}{(int, default = 5) Random seed for future replication.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) The channels to process.}

\item{...}{additional parameters to \code{\link[RcppAlgos:combinatoricsGeneral]{RcppAlgos::permuteGeneral()}} aside from n, m, Parallel and repetition}
}
\value{
e experiment object. The experiment object now has a list called \code{permutation_p_matrix} stored in it. Elements of this \code{permutation_p_matrix} list are
the outputs of different permutation comparison analyses. These elements are named by the groups that were compared.
}
\description{
This function performs a permutation analysis to compare the region pairwise correlation coefficients between two different analysis groups
specified by \code{correlation_list_name_1} and \code{correlation_list_name_2.} Note that both of these analysis groups must have the same
number of channels to compare. The functions \code{get_correlations()} needs to have been run for each of these analysis groups prior to
running this function.
}
