% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{create_perm_diff_network}
\alias{create_perm_diff_network}
\title{Create a permutation difference network
This function requires \code{\link[=export_permutation_results]{export_permutation_results()}} to have been run for the specific permutation analysis of interest, with the \code{filter_significant}
parameter set to FALSE, so that users can set their own p-value threshold for filtering out edges.}
\usage{
create_perm_diff_network(
  e,
  permutation_group,
  channel = "cfos",
  p_value_thresh = 0.001,
  ontology = "allen",
  anatomical.order = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB", "HB",
    "CB"),
  community_detection_function = tidygraph::group_fast_greedy,
  ...
)
}
\arguments{
\item{e}{experiment object}

\item{permutation_group}{(str) Supply names of the specific analyses found under \code{e$permutation_p_matrix \%>\% names()}}

\item{channel}{(str, default = "cfos")}

\item{p_value_thresh}{(numeric, default = 0.001) User chosen p-value for thresholding edges to include in the network.}

\item{ontology}{(str, default = "allen") Set to "unified" for Kim Lab unified ontology}

\item{anatomical.order}{(vec, default = c("Isocortex","OLF","HPF","CTXsp","CNU","TH","HY","MB","HB","CB"))
The default super region acronym list that groups all subregions in the dataset.}

\item{community_detection_function}{(function, default = tidygraph::group_fast_greedy)
\href{https://tidygraph.data-imaginist.com/reference/group_graph.html}{tidygraph grouping function} supporting
various algorithms for community detection}

\item{...}{Additional arguments for the community_detection_function chosen}
}
\value{
e experiment object
}
\description{
Create a permutation difference network
This function requires \code{\link[=export_permutation_results]{export_permutation_results()}} to have been run for the specific permutation analysis of interest, with the \code{filter_significant}
parameter set to FALSE, so that users can set their own p-value threshold for filtering out edges.
}
\examples{
\dontrun{
e <- create_perm_diff_network(e, permutation_group = "Control_vs_Isoflurane", channel = "cfos", ontology = "unified", p_value_thresh = 0.001, weights = NULL)
}

}
