% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_functions.R
\name{detect_single_slice_regions}
\alias{detect_single_slice_regions}
\title{Detect atlas regions that only show up in a single slice object}
\usage{
detect_single_slice_regions(m, remove = FALSE, output = TRUE)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Quality check function to make sure that the regions included for analysis show up
in more than 1 slice object, otherwise the user can remove them from the mouse object.

The user should run \code{\link[=normalize_cell_counts]{normalize_cell_counts()}} and \code{\link[=get_cell_table]{get_cell_table()}} functions prior to using this function.
If the user has run \code{\link[=split_hipp_DV]{split_hipp_DV()}} the function will automatically account for this
}
