% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exclude_anatomy.mouse}
\alias{exclude_anatomy.mouse}
\title{Method for excluding regions, layer 1, and hemispheres per mouse}
\usage{
\method{exclude_anatomy}{mouse}(
  m,
  slice_ID = "1_10",
  hemisphere = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  clean = TRUE,
  exclude_regions = NULL,
  exclude_other_hemisphere = TRUE,
  exclude_layer_1 = TRUE
)
}
\arguments{
\item{m}{: mouse object}

\item{slice_ID}{: ID of slice (str)}

\item{hemisphere}{: hemisphere of slice}

\item{channels}{: channels to filter (str vector), can filter both eyfp & cfos}

\item{clean}{clean: TRUE (bool). Remove cells that don't map to any regions.}

\item{exclude_regions}{: NULL (str vector); acronyms of regions you want to exclude IN ADDITION to regions that will by default be excluded in the slice attribute 'regions_excluded'}

\item{exclude_other_hemisphere}{: TRUE (bool); excludes the contralateral hemisphere from one indicated in slice attribute}

\item{exclude_layer_1}{: TRUE (bool); excludes all counts from layer 1 (TEMPORARY, may not be hardcoded in later)}
}
\value{
m mouse object
}
\description{
Method for excluding cell counts in specified regions, layer 1, out-of-bounds cells counts, and hemispheres per mouse.
PLEASE NOTE that this function automatically excludes the default regions included in the attribute "regions_excluded" in each slice
IN ADDITION to the regions added to the 'exclude_regions' parameter.
Regions added to the 'exclude_regions' parameter will then be updated in the slice attribute to keep track of overall which regions were excluded per slice.
}
\examples{
m <- exclude
}
