% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exclude_anatomy.slice}
\alias{exclude_anatomy.slice}
\title{Post-processing to exclude and clean up mapped data}
\usage{
\method{exclude_anatomy}{slice}(
  s,
  channels = c("cfos", "eyfp", "colabel"),
  clean = TRUE,
  exclude_regions = NULL,
  exclude_hemisphere = TRUE,
  exclude_layer_1 = TRUE,
  plot_filtered = TRUE
)
}
\arguments{
\item{s:}{slice object}

\item{channels:}{channels to filter (str vector), can filter both eyfp & cfos}

\item{clean:}{TRUE (bool). Remove cells that don't map to any regions.}

\item{exclude_regions:}{NULL (str vector); acronyms of regions you want to exclude IN ADDITION to regions that will by default be excluded in the slice attribute 'regions_excluded'}

\item{exclude_hemisphere:}{TRUE (bool); excludes the contralateral hemisphere from one indicated in slice attribute}

\item{exclude_layer_1:}{TRUE (bool); excludes all counts from layer 1 (TEMPORARY, may not be hardcoded in later)}
}
\description{
Method for excluding regions, layer 1, and hemispheres, and  per slice
}
