% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_functions.R
\name{exclude_by_acronym}
\alias{exclude_by_acronym}
\title{Excluded user chosen regions by entering acronyms}
\usage{
exclude_by_acronym(
  e,
  acronyms = "fiber_tracts",
  ontology = "allen",
  channels = NULL
)
}
\arguments{
\item{e}{experiment object}

\item{acronyms}{(str, default = "fiber_tracts") vector of region/structure acronyms to exclude from the datasets, e.g. c("CBL", "sox", "3V")}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}

\item{channels}{(str, default = NULL) NULL option processes all channels. \code{channels = c("cfos", "eyfp")} specifies exact channels.}
}
\value{
e experiment object
}
\description{
Excluded user chosen regions by entering acronyms
}
