% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_functions.R
\name{exclude_by_keyword}
\alias{exclude_by_keyword}
\title{Excluded user chosen regions by keywords found in long-form name}
\usage{
exclude_by_keyword(e, keywords, channels = NULL)
}
\arguments{
\item{e}{experiment object}

\item{keywords}{(str) vector of region/structure keywords to match and exclude from the datasets, e.g. c("nerve", "tract")}

\item{channels}{(str, default = NULL) NULL option processes all channels. \code{channels = c("cfos", "eyfp")} specifies exact channels.}
}
\value{

}
\description{
Excluded user chosen regions by keywords found in long-form name
}
