% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_functions.R
\name{exclude_redundant_regions}
\alias{exclude_redundant_regions}
\title{Exclude redundant regions}
\usage{
exclude_redundant_regions(e, ontology = "allen", channels = NULL)
}
\arguments{
\item{e}{exoeriment object}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}

\item{channels}{(str, default = NULL) NULL option processes all channels. \code{channels = c("cfos", "eyfp")} specifies exact channels.}
}
\value{
e experiment object
}
\description{
Your dataset may contain redundant information because it includes counts at different "ontological" resolutions.
SMARTR initially operates at the highest "resolution" and later on, can fold sub-regions into parent regions later.
Therefore redundant counts from parent regions should be removed from datasets using this function prior to analysis and plotting functions.
}
\details{
To check and see redundant regions contained in a list of acronyms, see the function \code{\link[=check_redundant_parents]{check_redundant_parents()}}.
}
