% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_object_constructors.R
\name{experiment}
\alias{experiment}
\title{Create an experiment object}
\usage{
my_experiment <- experiment() # construct an experiment object
}
\arguments{
\item{experiment_name}{(str, default = NULL)}

\item{experimenters}{(str, default = NULL)}

\item{experiment_groups}{(str, default = NULL) Autogenerated with \code{\link[=add_mouse]{add_mouse()}} function. Must exactly match the string values from the mouse objects.}

\item{drug_groups}{(str, default = NULL) Autogenerated with \code{\link[=add_mouse]{add_mouse()}} function. Must exactly match the string values from the mouse objects.}

\item{sex_groups}{(str, default = NULL) Autogenerated with \code{\link[=add_mouse]{add_mouse()}} function. Must exactly match the string values from the mouse objects.}

\item{cohorts}{(str, default = NULL) Autogenerated with \code{\link[=add_mouse]{add_mouse()}} function. Must exactly match the string values from the mouse objects.}

\item{strains}{(str, default = NULL) Autogenerated with \code{\link[=add_mouse]{add_mouse()}} function. Must exactly match the string values from the mouse objects.}

\item{genotypes}{(str, default = NULL) Autogenerated with \code{\link[=add_mouse]{add_mouse()}} function. Must exactly match the string values from the mouse objects.}

\item{reporters}{(str, default = NULL) Autogenerated with \code{\link[=add_mouse]{add_mouse()}} function. Must exactly match the string values from the mouse objects.}

\item{ages}{(str, default = NULL). AUtogenerated with \code{\link[=add_mouse]{add_mouse()}} function. Must exactly match the string values from the mouse objects.}

\item{output_path}{(str, default = 'set output path for your experiment') Where to save the RData file for your experiment object}

\item{...}{additional custom keyword pair attributes you'd like to store}
}
\value{
An experiment, a colloquial term for an object of class 'wb_experiment'. An 'experiment' object
is also a list, with class list.
}
\description{
experiment() constructs an S3 object of class 'wb_experiment'. An experiment object
consists of a list of processed mouse objects with raw data from slices omitted, and experimental attributes stored as a list.
}
\details{
The experimental attributes can be assigned as arguments to the experiment constructor function. See the parameters listed for the default values for these attributes
Note that you are able to add custom attributes as keyword pairs, if you would like to keep track of an additional piece of information. However, this will only serve a descriptive purpose
and will not be used for analysis. You may not need to use all experimental attributes but fill out as many are applicable to your experiment.
}
\seealso{
See also \code{\link[=mouse]{mouse()}} for the description of a mouse object and it's attributes.
}
