% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{export_permutation_results}
\alias{export_permutation_results}
\title{Export the permutation results as a csv file. This automatically saves into the tables folder.}
\usage{
export_permutation_results(
  e,
  permutation_groups = "all",
  channels = c("cfos"),
  ontology = "allen",
  filter_significant = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{permutation_groups}{(str vec, default = "all") Default is to export all analyses run. Supply names of the specific analyses found under \code{e$permutation_p_matrix \%>\% names()} if you want to export specific groups.}

\item{channels}{(str, channels =  c("cfos", "eyfp", "colabel") The channels to process.}

\item{ontology}{(str, default = "allen") Set to "unified" for Kim Lab unified ontology}

\item{filter_significant}{(bool, default = TRUE) If FALSE, this keeps all comparisons. Otherwise exports only the most different and significant permutations.}
}
\description{
Export the permutation results as a csv file. This automatically saves into the tables folder.
}
\examples{
\dontrun{
e <- export_permutation_results(e, permutation_groups = "all", filter_significant = TRUE)
}

}
