% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_functions.R
\name{filter_regions}
\alias{filter_regions}
\title{Filters to chosen base parent regions and all child subregions}
\usage{
filter_regions(
  e,
  base_regions = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB", "HB",
    "CBL"),
  ontology = "allen",
  channels = NULL
)
}
\arguments{
\item{e}{experiment object}

\item{base_regions}{(default = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU","TH", "HY", "MB", "HB", "CBL"))
Region acronyms of base parent regions (and all subregions) to include for analysis}

\item{ontology}{(str, default = "allen") Options = "allen" or "unified".}

\item{channels}{(str, default = NULL) NULL option processes all channels. \code{channels = c("cfos", "eyfp")} specifies exact channels.}
}
\value{
e experiment object
}
\description{
Filters to chosen base parent regions and all child subregions
}
