% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{find_all_subregions}
\alias{find_all_subregions}
\title{find_all_subregions}
\usage{
find_all_subregions(regions)
}
\arguments{
\item{regions}{A string vector of the Allen Mouse Brain Atlas abbreviated}
}
\value{
all_regions. A string vector of all the input & subregions within them to exclude
}
\description{
find_all_subregions
}
