% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{find_excluded_subregions}
\alias{find_excluded_subregions}
\title{find_excluded_subregions}
\usage{
find_excluded_subregions(exclude_regions)
}
\arguments{
\item{excluded_regions}{A string vector of the Allen Mouse Brain Atlas abbreviated
regions of regions to exclude}
}
\value{
all_excluded_regions. A string vector of all the input & subregions within them to exclude
}
\description{
find_excluded_subregions
}
