% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_functions.R
\name{find_outlier_counts}
\alias{find_outlier_counts}
\title{Remove outlier counts}
\usage{
find_outlier_counts(
  e,
  by = c("group", "sex"),
  n_sd = 2,
  remove = FALSE,
  log = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{by}{(str, default = c("group", "sex")) The mice attributes used to group the datasets into comparison groups.}

\item{n_sd}{(int, default = 2). Number of standards deviations above and below which categorizes outliers.}

\item{remove}{(bool, default = FALSE) Remove all outlier rows from the combined normalized counts dataframe in the experiment object.}

\item{log}{(bool, default = TRUE) Save the logged outlier regions into a csv file in the output folder.}
}
\value{
e experiment object. Outlier counts in the experiment object are removed if remove = TRUE.
}
\description{
remove any normalized regions counts that are more than \code{n_sd} standard deviations (default = 2) higher
}
\examples{
e <- find_outlier_counts(e, by = c("group","sex"), n_sd = 2, remove = FALSE, log = TRUE)

}
