% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{get.acronym.child}
\alias{get.acronym.child}
\title{Get acronyms of child structures}
\usage{
get.acronym.child(x)
}
\arguments{
\item{x}{(str) Regional acronym}
}
\description{
Function to get the acronym of parent regions in the ontology
}
