% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{get.acronym.parent.custom}
\alias{get.acronym.parent.custom}
\title{Get parent region acronyms}
\usage{
get.acronym.parent.custom(x, ontology = "unified")
}
\arguments{
\item{x}{(str) Regional acronym}

\item{ontology}{(str, options: "unified")}
}
\description{
Function to get the acronym of parent regions
}
