% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.acronym.parent}
\alias{get.acronym.parent}
\title{Get parent region acronyms}
\usage{
get.acronym.parent(x)
}
\arguments{
\item{x}{(str) Regional acronym}
}
\description{
Function to get the acronym of parent regions
}
