% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.registered.areas}
\alias{get.registered.areas}
\title{Title}
\usage{
get.registered.areas(cell.data.list, registration, conversion.factor = 1)
}
\arguments{
\item{conversion.factor}{}
}
\value{

}
\description{
Title
}
