% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.sub.structure.custom}
\alias{get.sub.structure.custom}
\title{Get subregion acronyms}
\usage{
get.sub.structure.custom(x, ontology = "unified")
}
\arguments{
\item{x}{(str vector) Regional acronyms in a vector}

\item{ontology}{(str, options: "unified")}
}
\description{
Search function to get ALL substructure acronyms from parent acronym. Note this functions is recursive.
}
