% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.sub.structure}
\alias{get.sub.structure}
\title{Get subregion acronyms}
\usage{
get.sub.structure(x)
}
\arguments{
\item{x}{(str vector) Regional acronyms in a vector}
}
\description{
Search function to get ALL substructure acronyms from parent acronym.
}
