% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.sup.structure.custom}
\alias{get.sup.structure.custom}
\title{Get super parent region acronyms}
\usage{
get.sup.structure.custom(
  x,
  ontology = "unified",
  matching.string = c("root", "grey", "CH", "VS", "CTX", "CNU", "MB", "HB")
)
}
\arguments{
\item{x}{(str) Regional acronym}

\item{ontology}{(str, options: "unified")}

\item{matching.string}{(str vector, default = c("CTX", "CNU", "IB", "MB", "HB", "grey", "root", "VS", "fiber tracts")) Vector of the basest parent levels to stop at.}
}
\description{
Function to get the acronym of super parent regions in the ontology
}
