% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.sup.structure}
\alias{get.sup.structure}
\title{Get parent region acronyms}
\usage{
get.sup.structure(
  x,
  matching.string = c("CTX", "CNU", "IB", "MB", "HB", "grey", "root", "VS",
    "fiber tracts")
)
}
\arguments{
\item{x}{(str) Regional acronym}

\item{matching.string}{(str vector, default = c("CTX", "CNU", "IB", "MB", "HB", "grey", "root", "VS", "fiber tracts")) Vector of the basest parent levels to stop at.}
}
\description{
Function to get the acronym of parent regions in the ontology
}
