% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{get.super.regions}
\alias{get.super.regions}
\title{Updated function to return character vector of the super region acronyms after inputting a character vector of acronyms}
\usage{
get.super.regions(
  acronym,
  anatomical.order = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB", "HB",
    "CB"),
  ontology = "allen"
)
}
\arguments{
\item{acronym}{(str vec) character vector of acronyms. Cannot be a factor vector.}

\item{anatomical.order}{(default = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU","TH", "HY", "MB", "HB", "CB")) Default way to group subregions into super regions order}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}
}
\value{
vector the same length as acronym consisting of the super regions of the acronyms
}
\description{
Updated function to return character vector of the super region acronyms after inputting a character vector of acronyms
}
\examples{
get.super.regions(acronym = c("DG-sg", "ACA5", "CLA", "ACB", "BLA", "PERI" , "PB"))
get.super.regions(acronym = c("CPi", "MHb", "ZI", "Acb", "BL"), ontology = "unified")
}
