% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{get_cell_table}
\alias{get_cell_table}
\title{Get cell tables}
\usage{
get_cell_table(m, channels = c("cfos", "eyfp", "colabel"))
}
\arguments{
\item{m}{mouse object}

\item{channels}{(vec) Channels to process.}
}
\value{
m mouse object
}
\description{
This function stores a list in a mouse object that is the length of the channels parameter. Each element of the list is a dataframe
containing combined cell counts for one channel across all slices processed for that mouse.
By default, if one slice has no dataset processed for that particular channel, that slice will be skipped over.
The function will run properly but a warning will be thrown indicating you should go back and
generate a mapped dataset for that particular slice and channel.
}
\examples{
m <- get_cell_tables(m, channels = c("cfos", "eyfp", "colabel"))
}
