% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{get_hipp_DV_volumes}
\alias{get_hipp_DV_volumes}
\title{Get aggregate volumes in the hippocampus}
\usage{
get_hipp_DV_volumes(
  m,
  AP_coord = -2.7,
  rois = c("DG", "CA1", "CA2", "CA3"),
  hipp_AP_coordinates
)
}
\arguments{
\item{AP_coord}{}
}
\value{
a list the length of channels, total_volumes
}
\description{
Get aggregate volumes in the hippocampus
}
