% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{get_percent_colabel}
\alias{get_percent_colabel}
\title{Get the percentage of colabelled cells over cfos or eyfp channels}
\usage{
get_percent_colabel(e, channel = "cfos", save_table = TRUE, roi = NULL)
}
\arguments{
\item{channel}{(str, default = "cfos") The channel used as denominator in percentage counts.}

\item{save_table}{(bool, default = TRUE) Whether to save the output table as a csv in the experiment object output folder.}

\item{roi}{(str, default = NULL) Whether to generate colabelled percentages for only specific regions of interest. The default is
to assess across all rois.}
}
\value{
e experiment object with colabel percent table
}
\description{
This analysis will only include common regions that are included in both the colabelled and
cfos or eyfp channels. The colabelled percentage of individual animals will be extracted. These can be
}
