% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{get_registered_volumes.mouse}
\alias{get_registered_volumes.mouse}
\title{Method for getting regional areas and volumes for each slice in a mouse object}
\usage{
\method{get_registered_volumes}{mouse}(m, slice_ID, hemisphere = NULL, replace = FALSE)
}
\arguments{
\item{m}{mouse object}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str) 'left', 'right' or NULL (both)}
}
\value{
m mouse object
}
\description{
Method for getting regional areas and volumes for each slice in a mouse object
}
\examples{
m <- get_registered_areas(m, slice_ID = "1_10", hemisphere = "left", replace = FALSE)
}
