% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_registered_areas.slice}
\alias{get_registered_areas.slice}
\title{Method for getting regional areas per slice}
\usage{
\method{get_registered_areas}{slice}(s)
}
\arguments{
\item{s}{: slice object}
}
\value{
s slice object
}
\description{
returns a dataframe with columns 'name' (full region name), 'acronym', 'area', 'right.hemisphere
}
\examples{
s <- get_registered_areas(2)
}
