% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id.from.acronym.custom}
\alias{id.from.acronym.custom}
\title{Get region ontology  ID from acronym}
\usage{
id.from.acronym.custom(x, ontology = "unified")
}
\arguments{
\item{x}{(int) integer ID}

\item{ontology}{(str, options: "unified")}
}
\description{
Similar to wholebrain package's search functions to get whole regional names from a numerical ID lookup table. For custom ontologies.
}
