% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_mapped_datasets}
\alias{import_mapped_datasets}
\title{Import externally mapped datasets into an experiment}
\usage{
import_mapped_datasets(e, normalized_count_paths, ...)
}
\arguments{
\item{e}{experiment object}

\item{normalized_count_paths}{(str vec, default = NULL, optional) For importing external datasets ONLY.
A character vector, with each element being a path to the respective .csv or .xlsx file in the same order as the \code{channels} attribute.
Must have the same order and length as the \code{channels} parameter.
You can add channels as names to the vector elements to avoid ambiguity and ensure correct importation for the right channel.}

\item{...}{additional parameters to pass to either the \code{\link[readr:read_delim]{readr::read_csv()}} function or \code{\link[readxl:read_excel]{readxl::read_excel()}}}
}
\value{
e an experiment object with the imported dataset
}
\description{
This function takes an experiment object and imports externally mapped datasets.
}
