% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_segmentation.mouse}
\alias{import_segmentation.mouse}
\title{Import raw segmentation data}
\usage{
\method{import_segmentation}{mouse}(
  m,
  slice_ID = "1_10",
  hemisphere = NULL,
  channels = c("cfos", "efyp", "colabel"),
  replace = FALSE
)
}
\arguments{
\item{m}{: mouse object}

\item{slice_ID}{: ID of slice (str)}

\item{hemisphere}{: 'left', 'right' or NULL (both)}

\item{channels}{: channels to import (str vector)}

\item{replace}{: replace existing raw segmentation data (bool)}
}
\value{
m mouse object
}
\description{
Method for importing segmentation data for a mouse object
}
