% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{import_segmentation}
\alias{import_segmentation}
\alias{import_segmentation.slice}
\alias{import_segmentation.mouse}
\title{import_segmentation (generic function)}
\usage{
import_segmentation(x, ...)

\method{import_segmentation}{slice}(s, mouse_ID = NA, channels = c("cfos", "eyfp", "colabel"))

\method{import_segmentation}{mouse}(
  m,
  slice_ID = NA,
  hemisphere = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  replace = FALSE
)
}
\arguments{
\item{...}{}

\item{s}{slice object}

\item{mouse_ID}{(str) ID of mouse}

\item{channels}{(str vector) channels to import}

\item{m}{mouse object}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str)'left', 'right' or NULL}

\item{replace}{(bool, default = FALSE) replace existing raw segmentation data}
}
\value{


s slice object

m mouse object
}
\description{
Method for importing segmentation data for a slice object

Method for importing segmentation data for a mouse object
}
\note{
Note that in order to import the raw data for co-labeled cells, you need both the txt files that end in "ColocOnly.txt" and "eYFP_16bit.txt".
The ImageJ colabelling algorithm calculates overlap between two segmented 3D objects from different channel.
One channel serves as the "reference" to later generate X & Y coordinates of colabelled cells in the make_segmentation_object() function.
By convention, we use the eYFP channel rather than cfos..
}
\examples{
s <-  import_segmentation(s, mouse_ID = "255", channels = c("cfos", "eyfp", "colabel"))
m <-  import_segmentation(m, slice_ID = "1_10", channels = c("cfos", "eyfp", "colabel"), replace = FALSE)
}
