% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_segmentation.slice}
\alias{import_segmentation.slice}
\title{Import segmentation data}
\usage{
\method{import_segmentation}{slice}(
  s,
  mouse_ID = "255",
  channels = c("cfos", "eyfp", "colabel")
)
}
\arguments{
\item{s}{: slice object}

\item{mouse_ID}{: ID of mouse (str)}

\item{channels}{: channels to import (str vector)}
}
\value{
s slice object
}
\description{
Method for importing segmentation data for a slice object
}
\examples{
s <-  import_segmentation(s, mouse_ID = "255", channels = c("cfos", "eyfp", "colabel"))
TODO: modify paths to search for correct files regardless of upper or lower case. Use grep and stringi to match multiple patterns
}
