% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{import_segmentation_custom}
\alias{import_segmentation_custom}
\alias{import_segmentation_custom.slice}
\alias{import_segmentation_custom.mouse}
\title{import_segmentation (generic function)}
\usage{
import_segmentation_custom(x, ...)

\method{import_segmentation_custom}{slice}(
  s,
  channel,
  x_col = NULL,
  y_col = NULL,
  meas_path = NULL,
  quant_path = NULL
)

\method{import_segmentation_custom}{mouse}(
  m,
  channel,
  slice_ID = NA,
  hemisphere = NULL,
  x_col = NULL,
  y_col = NULL,
  meas_path = NULL,
  quant_path = NULL
)
}
\arguments{
\item{...}{}

\item{s}{slice object}

\item{channel}{(str) channel to import}

\item{x_col}{= NULL,}

\item{y_col}{= NULL,}

\item{meas_path}{= NULL,}

\item{quant_path}{= NULL}

\item{m}{mouse object}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str)'left', 'right' or NULL}
}
\value{


s slice object

m mouse object
}
\description{
Custom method for importing segmentation data for a slice object.
This is a flexible method for importing channels aside from \code{cfos} and \code{eyfp} that use the same ImageJ segmentation macros common to the
Denny Lab. Other labs may use this method. This method works best when following saving segmentation data with the same naming conventions using the
\href{https://imagejdocu.list.lu/plugin/stacks/3d_ij_suite/start}{3D Roi Manager} within the
\href{dx.doi.org/10.1093/bioinformatics/btt276}{3D ImageJ Suite}.
Segmentation data is saved in two'.txt' files which output the results of the Measure 3D and Quantif 3D options of the 3D Roi Manager plugin,respectively.
A \href{https://imagejdocu.list.lu/tutorial/working/tutorial_for_3d_roi_manager}{description} of what each of those these options measure
is provided in the online documentation.

The naming conventions for the ".txt" file storing the Quantif3D results are \verb{Q_*_\{channel\}_*_\{channel\}.txt},
where the \code{*} indicates a wildcard character(s) and {channel} is the channel name without brackets.
E.g. "Q_G_eYFP_258_1_1_eYFP.txt"

The naming conventions for the ".txt" file storing the Measure 3D are \verb{M_*_\{channel\}_*.txt} where
the \code{*} indicates a wildcard character(s) and {channel} is the channel name without brackets.
E.g "M_G_eYFP_258_1_1.txt".

The wildcards characters may be used to store things like date or slice naming information.

The locations of these files must be specified in the \code{slice_directory} attribute of the slice_object.
Otherwise, the root folder containing the registration image is searched.
This attribute can be stored when initializing the slice object or can be edited afterwards.

Method for custom importation of segmentation data for a mouse object
}
\examples{
s <-  import_segmentation_custom(s, mouse_ID = "255", channel = "cfos")
m <-  import_segmentation(m, slice_ID = "1_10", channels = c("PV"), replace = FALSE)
}
