% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{import_segmentation_ij}
\alias{import_segmentation_ij}
\alias{import_segmentation_ij.slice}
\alias{import_segmentation_ij.mouse}
\title{import_segmentation (generic function)}
\usage{
import_segmentation_ij(x, ...)

\method{import_segmentation_ij}{slice}(s, mouse_ID = NA, channels = NULL)

\method{import_segmentation_ij}{mouse}(
  m,
  slice_ID = NA,
  hemisphere = NULL,
  channels = NULL,
  replace = FALSE
)
}
\arguments{
\item{...}{}

\item{s}{slice object}

\item{mouse_ID}{(str) ID of mouse}

\item{channels}{(str vector, default = NULL) channels to import. If NULL, defaults to the channels stored in the slice object attributes.}

\item{m}{mouse object}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str)'left', 'right' or NULL}

\item{replace}{(bool, default = FALSE) replace existing raw segmentation data}
}
\value{


s slice object

m mouse object
}
\description{
Method for importing segmentation data for a slice object

Method for importing segmentation data for a mouse object
}
\note{
The designated \code{colabel} channel name in this pipeline refers to the overlap between \code{cfos} and \code{eyfp} channels.
To name another channel for colocalized cells between two other channels, please use a different naming convention,
e.g. "colabel_PV_cfos" and import using a customized import function such as \code{\link[=import_segmentation_custom]{import_segmentation_custom()}}.
Note that in order to import the raw data for \code{eyfp} and \code{cfos} co-labeled cells, you need both the
txt files that end in "ColocOnly.txt" and "eYFP_16bit.txt". The ImageJ colabelling algorithm calculates overlap
between two segmented 3D objects from different channel.
One channel serves as the "reference" to later generate X & Y coordinates of colabelled cells in the make_segmentation_object() function.
By convention, we use the eYFP channel rather than cfos..
}
\examples{
s <-  import_segmentation(s, mouse_ID = "255") # Defaults to channels stored in slice attributes
s <-  import_segmentation(s, mouse_ID = "255", channels = c("cfos", "eyfp", "colabel")) # Specify channels
m <-  import_segmentation(m, slice_ID = "1_10", channels = c("cfos", "eyfp", "colabel"), replace = FALSE)
}
