% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_segmentation_object.mouse}
\alias{make_segmentation_object.mouse}
\title{Creates a wholebrain segmentation object and stores it}
\usage{
\method{make_segmentation_object}{mouse}(
  m,
  slice_ID = "1_10",
  hemisphere = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  replace = FALSE,
  use_filter = TRUE
)
}
\arguments{
\item{m}{: mouse object}

\item{slice_ID}{: ID of slice (str)}

\item{channels}{: (vec) Channels to process; Note: always set 'eyfp' before 'colabel' in the vector}

\item{replace}{: replace existing raw segmentation data (bool)}

\item{hemisphere:}{'left', 'right' or NULL (both)}

\item{use_filter:}{TRUE (bool). Use a filter to create more curated segmentation object from the raw segmentation data.}
}
\description{
Make a wholebrain compatible segmentation object for a slice in a mouse object
}
