% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_segmentation_object}
\alias{make_segmentation_object}
\title{make_segmentation_object}
\usage{
make_segmentation_object(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
make_segmentation_object
}
