% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_segmentation_object.slice}
\alias{make_segmentation_object.slice}
\title{Create a segmentation object
Creates a wholebrain segmentation object and stores it.}
\usage{
\method{make_segmentation_object}{slice}(
  s,
  mouse_ID = "255",
  channels = c("cfos", "eyfp", "colabel"),
  use_filter = TRUE
)
}
\arguments{
\item{s}{: slice object}

\item{mouse_ID}{: ID of mouse (str)}

\item{channels}{: (str vec) Channels to process; Note: always set 'eyfp' before 'colabel' in the vector}

\item{use_filter}{: TRUE (bool). Use a filter to create more curated segmentation object from the raw segmentation data.}
}
\value{
s slice object
}
\description{
Make a wholebrain compatible segmentation object for a slice in a slice object
}
\examples{

}
