% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_cells_to_atlas.mouse}
\alias{map_cells_to_atlas.mouse}
\title{Map cells to atlas for slice object}
\usage{
\method{map_cells_to_atlas}{mouse}(
  m,
  slice_ID = "1_10",
  hemisphere = "left",
  channels = c("cfos", "efyp", "colabel"),
  clean = TRUE,
  display = TRUE,
  replace = FALSE,
  ...
)
}
\arguments{
\item{m}{: mouse object}

\item{slice_ID}{: ID of slice (str)}

\item{hemisphere}{'left', 'right' or NULL (both)}

\item{channels}{(vec of str) Channels to process; Note: always set 'eyfp' before 'colabel' in the vector}

\item{clean}{: TRUE (bool). Remove cells that don't map to any regions.}

\item{display}{: TRUE (bool). Display the results of the forward warp for the slice.display}

\item{replace}{: FALSE (bool). Replace current forward warped data, both raw and cleaned.}

\item{...}{additional parameters besides 'registration', 'segmentation', 'forward.warps', and 'device' to pass to the wholebrain::inspect.registration() function}
}
\value{
m mouse object
}
\description{
Map cells to atlas for slice object
}
\examples{

}
