% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{map_cells_to_atlas.slice}
\alias{map_cells_to_atlas.slice}
\title{Map cells to atlas for slice object}
\usage{
\method{map_cells_to_atlas}{slice}(
  s,
  channels = c("cfos", "eyfp", "colabel"),
  clean = TRUE,
  display = TRUE,
  ...
)
}
\arguments{
\item{s}{: slice object}

\item{channels}{: channels to filter (str vector), can filter both eyfp & cfos}

\item{clean}{: TRUE (bool). Remove cells that don't map to any regions.}

\item{display}{: TRUE (bool). Display the results of the forward warp for the slice.}

\item{...}{additional parameters besides 'registration', 'segmentation', 'forward.warps', and 'device' to pass to the wholebrain::inspect.registration() function}
}
\description{
Method for forward warping segmentation data to atlas space for a slice object.
Requires segmentation objects to be made for channels specified and a registration object.
}
