% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{maslov_sneppen_rewire}
\alias{maslov_sneppen_rewire}
\title{Internal algorithm for maslov-sneppen rewiring}
\usage{
maslov_sneppen_rewire(network, n_rewires = 10000)
}
\arguments{
\item{network}{tidygraph graph object}

\item{n_rewires}{number of rewires.}
}
\value{
a tidygraph graph
}
\description{
This algorithm is not appropriate if you would like to take weights into account
See Maslov & Sneppen (2002)"Specificity and stability in topology of protein networks"
}
