% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{name.from.acronym.custom}
\alias{name.from.acronym.custom}
\title{Get region ontology name from acronym}
\usage{
name.from.acronym.custom(x, ontology = "unified")
}
\arguments{
\item{x}{(str) Regional acronym or vector of regional acronyms}

\item{ontology}{(str, options: "unified")}
}
\description{
Get whole regional names from acronyms based on a custom lookup table imported to accomodate other ontologies.
}
