% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{name.from.acronym}
\alias{name.from.acronym}
\title{Get region ontology name from acronym}
\usage{
name.from.acronym(x)
}
\arguments{
\item{x}{(str) Regional acronym}
}
\description{
Get whole regional names from acronyms based on a lookup table including SMARTR's custom ontology for the
dorsal ventral split of the hippocampus
}
