% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{name.from.id.custom}
\alias{name.from.id.custom}
\title{Get region ontology name from ID}
\usage{
name.from.id.custom(x, ontology = "unified")
}
\arguments{
\item{x}{(int) integer ID}

\item{ontology}{(str, options: "unified")}
}
\description{
Similar to wholebrain package's search functions to get whole regional names from a numerical ID lookup table. For custom ontologies.
}
