% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name.from.id}
\alias{name.from.id}
\title{Get region ontology name from ID}
\usage{
name.from.id(x)
}
\arguments{
\item{x}{(int) integer ID}
}
\description{
Similar to wholebrain package's search functions to get whole regional names from a numerical ID lookup table
including SMARTR's custom ontology for the  dorsal ventral split of the hippocampus
}
