% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize.registered.areas}
\alias{normalize.registered.areas}
\title{normalize registered areas}
\usage{
normalize.registered.areas(cell_table, total_areas, z_width = 24)
}
\arguments{
\item{cell_table}{cell_tables list from get_cell_table()}

\item{total_areas}{computed total areas list}

\item{z_width}{stack width in microns (default = 24)}
}
\value{

}
\description{
normalize registered areas
}
