% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{normalize_cell_counts}
\alias{normalize_cell_counts}
\title{Normalize cell counts per mm^2 or by mm^3 (if multiplying by the stack size)}
\usage{
normalize_cell_counts(m, combine_hemispheres = TRUE, simplify_regions = TRUE)
}
\arguments{
\item{simplify_regions}{}
}
\value{

}
\description{
Run this function after all the slices that you want to process are finished being added
and you have run the function get_cell_table() for your mouse.
}
\examples{
m <- normalize_cell_counts(m, combine_hemispheres = TRUE, simplify_regions = TRUE)
}
