% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R, R/utils.R
\name{normalize_colabel_counts}
\alias{normalize_colabel_counts}
\title{Normalize colabel counts over a designated denominator channel.}
\usage{
normalize_colabel_counts(e, denominator_channel = "eyfp")

normalize_colabel_counts(e, denominator_channel = "eyfp")
}
\arguments{
\item{e}{experiment object}

\item{denominator_channel}{(str, default = "eyfp") The exact name of the channel used for normalization}
}
\value{
e An experiment object with a new dataframe with the normalized ratios of colabelled counts over the designated denominator counts.
Because the volumes and region areas cancel out, the values of count, normalized.count.by.area, and normalized.count.by.volume are all the same.
This is to provide a consistent input dataframe into the analysis functions.

e An experiment object with a new dataframe with the normalized ratios of colabelled counts over the designated denominator counts.
Because the volumes and region areas cancel out, the values of count, normalized.count.by.area, and normalized.count.by.volume are all the same.
This is to provide a consistent input dataframe into the analysis functions.
}
\description{
This function can only be run after running \code{\link[=combine_norm_cell_counts]{combine_norm_cell_counts()}}. It divides the colabelled cell counts by
a designated normalization channel to provide a normalized ratio. Please note that the areas and volumes cancel out in this operation.
This is not designed to work on multiple hemispheres. Please combine cell counts across multiple hemispheres when you run \code{\link[=normalize_cell_counts]{normalize_cell_counts()}}.

This function can only be run after running \code{\link[=combine_norm_cell_counts]{combine_norm_cell_counts()}}. It divides the colabelled cell counts by
a designated normalization channel to provide a normalized ratio. Please note that the areas and volumes cancel out in this operation.
This is not designed to work on multiple hemispheres. Please combine cell counts across multiple hemispheres when you run \code{\link[=normalize_cell_counts]{normalize_cell_counts()}}.
}
\examples{
\dontrun{
e <- normalize_colabel_counts(e, denominator_channel = "eyfp")
}
e <- normalize_colabel_counts(e, denominator_channel = "eyfp")
}
\seealso{
\code{\link[=combine_norm_cell_counts]{combine_norm_cell_counts()}} &  \code{\link[=normalize_cell_counts]{normalize_cell_counts()}}

\code{\link[=combine_norm_cell_counts]{combine_norm_cell_counts()}} &  \code{\link[=normalize_cell_counts]{normalize_cell_counts()}}
}
