% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontology.unified}
\alias{ontology.unified}
\title{Unified Kim ontology}
\format{
A dataframe
\describe{
}
}
\usage{
ontology.unified
}
\description{
The combined ontology created by the Yongsoo Kim lab. Combines nomenclature of the two most-used mouse brain atlases, the Franklin and Paxinos (FP) and the common coordinate framework (CCF) from the Allen Institute for Brain Science.
}
\keyword{datasets}
