% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{parallel_coordinate_plot}
\alias{parallel_coordinate_plot}
\title{Create a parallel coordinate plot}
\usage{
parallel_coordinate_plot(
  e,
  permutation_comparison = "AD_vs_control",
  channels = c("cfos", "eyfp", "colabel"),
  colors = c("#be0000", "#00782e", "#f09b08"),
  x_label_group_1 = NULL,
  x_label_group_2 = NULL,
  height = 10,
  width = 10,
  print_plot = TRUE,
  save_plot = TRUE,
  image_ext = ".png"
)
}
\arguments{
\item{e}{experiment object}

\item{permutation_comparison}{The name of the correlation group comparisons to plot.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) channels to plot}

\item{colors}{(str, default = c("#be0000", "#00782e", "#f09b08")) Hexadecimal codes corresponding to the channels (respectively) to plot.}

\item{x_label_group_1}{(str, NULL) The label for the first group in the permutation analysis.}

\item{x_label_group_2}{(str, NULL) The label for the second group in the permutaiton analysis.}

\item{height}{height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{print_plot}{(bool, default = TRUE) Whether to display the plot (in addition to saving the plot)}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}

\item{image_ext}{(default = ".png") image extension to save the plot as.}
}
\description{
Plot the correlation difference between two comparison groups into a parallel coordinate plot. The function
\code{\link[=correlation_diff_permutation]{correlation_diff_permutation()}} must be run first in order to generate results to plot.
}
