% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parentid.from.id.custom}
\alias{parentid.from.id.custom}
\title{Get parent id from id}
\usage{
parentid.from.id.custom(x, ontology = "unified")
}
\arguments{
\item{x}{}

\item{ontology}{}
}
\description{
Get parent id from id
}
