% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{permute_corr_diff_distrib}
\alias{permute_corr_diff_distrib}
\title{Generate array of null distribution of region pairwise correlation differences.}
\usage{
permute_corr_diff_distrib(
  df,
  correlation_list_name_1,
  correlation_list_name_2,
  n_shuffle = n_shuffle,
  seed = 5,
  ...
)
}
\arguments{
\item{seed}{random seed for replication}
}
\value{

}
\description{
Generate array of null distribution of region pairwise correlation differences.
}
