% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_betweenness_regions}
\alias{plot_betweenness_regions}
\title{Plot the betweenness distributions across regions}
\usage{
plot_betweenness_regions(
  e,
  network,
  channels = c("cfos", "eyfp"),
  colors = c("red", "green"),
  title = "my_title",
  height = 10,
  width = 20,
  ylim = c(0, 60),
  image_ext = ".png",
  print_plot = TRUE,
  save_plot = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{network}{(str) Which network to plot the betweenness distribution across regions}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) Channels to plot}

\item{colors}{(str, default = ) String vector of hexadecimal color codes corresponding to to each channel plotted.}

\item{title}{(str, default = "my_title")
the experiment object output folder.}

\item{height}{(int, default = 15) Height of the plot in inches.}

\item{width}{(int, default = 20) Width of the plot in inches.}

\item{ylim}{(vec, default = c(0,15))axes limits of y-axis}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{print_plot}{(bool, default = TRUE) Whether to print the plot as an output.s}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Bar plot of betweenness per region in descending magnitude
}
