% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_cell_counts}
\alias{plot_cell_counts}
\title{This function allows for plotting of normalized cell counts by area across specific regions to plot.}
\usage{
plot_cell_counts(
  e,
  channel = "eyfp",
  rois = c("AAA", "dDG", "HY"),
  color_mapping = "group",
  colors = c("#952899", "#358a9c"),
  pattern_mapping = NULL,
  patterns = c("gray100", "hs_fdiagonal", "hs_horizontal", "gray90", "hs_vertical"),
  ylab = bquote(`Cell counts `("cells/mm"^3)),
  error_bar = "sem",
  ylim = c(0, 100),
  plot_individual = TRUE,
  height = 8,
  width = 8,
  print_plot = TRUE,
  save_plot = TRUE,
  image_ext = ".png"
)
}
\arguments{
\item{e}{experiment object}

\item{channel}{(str, default = "eyfp") The channel used as denominator in fraction counts.}

\item{rois}{(vec,  default = c("AAA", "dDG", "HY")) Allen acronyms of the ROIS that the user would like to plot}

\item{color_mapping}{(str, default = "group") The variable name that maps subgroups  you would like to graphically distinguish through colors.}

\item{colors}{(vec, default = c("#952899", "#358a9c")) A vector of hexadecimal color codes for each subgroup distinguished by the color mapping variable.}

\item{pattern_mapping}{(str, default = NULL) variable name that maps subgroups  you would like to graphically distinguish through bar patterns.Set to NULL if not in use.}

\item{patterns}{(default = c("gray100", 'hs_fdiagonal', "hs_horizontal", "gray90", "hs_vertical") Available patterns in ggpattern package to map to subgroups distinguished by the pattern mapping variable.}

\item{ylab}{(str, default =  bquote('Cell counts '('cells/mm'^3))) unit of measurement}

\item{error_bar}{(str, c("sd", "sem)) options for which type of error bar to display, standard deviation or standard error of the mean.}

\item{ylim}{(vec, default = c(0,100)) Y-axis range.}

\item{plot_individual}{(bool) whether to plot individual points}

\item{height}{(numeric) height of the plot in inches to save as.}

\item{width}{(numeric) height of the plot in inches to save as.}

\item{print_plot}{(bool) Whether to to print the plot.}

\item{save_plot}{(bool) Whether to save the plot in the experiment figures folder.}

\item{image_ext}{(str, default = ".png") Extension of the output file}
}
\value{
p Plot handle to the figure
}
\description{
Two different mouse attributes can be used as categorical variables to map to either the color or
pattern aesthetics of the bar plot, e.g. sex and experimental group.
The color aesthetic takes precedence over the pattern aesthetic so if you only want to use one mouse attribute, for plotting
set it to the \code{color_mapping} parameter and set the \code{pattern_mapping} parameter to NULL.
}
\examples{
plot_percentage_colabel
}
