% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_colabel_percent}
\alias{plot_colabel_percent}
\title{plot_colabel_percent}
\usage{
plot_colabel_percent(e, by = c("group", "sex"), roi = NULL)
}
\arguments{
\item{roi}{}
}
\value{

}
\description{
plot_colabel_percent
}
