% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_correlation_heatmaps}
\alias{plot_correlation_heatmaps}
\title{plot_colabel_percent}
\usage{
plot_correlation_heatmaps(
  e,
  colors = c("#be0000", "#00782e", "#0d7983"),
  save_plot = TRUE,
  title = NULL,
  height = 10,
  width = 10,
  image_ext = ".png"
)
}
\arguments{
\item{e}{experiment object that contains a correlation_list object generated by \code{\link[=get_correlations]{get_correlations()}}}

\item{colors}{Hexadecimal code for the colors corresponding to the channels attribute of the correlation_list.}

\item{save_plot}{(bool, default = TRUE) Save the correlation heatmap plots into the figures subdirectory of the
the experiment object output folder.}

\item{title}{Title of the plot. If NULL, the grouping variable names will be taken from the correlation_list object annd used as the title.}

\item{height}{Heigh of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{image_ext}{(default = ".png") image extension to the plot as=.}

\item{roi}{}
}
\value{


e experiment object
}
\description{
plot_colabel_percent
}
\examples{
Plot correlation heatmaps

}
