% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_degree_distributions}
\alias{plot_degree_distributions}
\title{Plot the degree distributions}
\usage{
plot_degree_distributions(
  e,
  channels = c("cfos", "eyfp"),
  color_palettes = c("reds", "greens"),
  colors_manual = NULL,
  labels = c(female_AD = "female_AD_label", female_control = "female_control_label"),
  title = "my_title",
  height = 15,
  width = 15,
  xlim = c(0, 20),
  ylim = c(0, 15),
  image_ext = ".png",
  print_plot = TRUE,
  save_plot = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) Channels to plot.}

\item{color_palettes}{(str, default = c("reds", "greens")) Color palettes from \link[grDevices:palettes]{grDevices::hcl.colors} that are used to for plotting networks for each channel, respectively.}

\item{colors_manual}{(str, default = NULL ) Manually choose the hexadecimal color codes to create a custom color palette, e.g. colors_manual = c("#660000", "#FF0000", "#FF6666").
Warning: this color will be applied to all channels. It's recommended to set the channels parameter to a single channel if this parameter is used.}

\item{labels}{(e.g. labels = c(network1_name = "network 1 label", network2_name = "network 2 label))
The legend labels to correspond with your respective network names.}

\item{title}{(str, default = "my_title")
the experiment object output folder.}

\item{height}{(int, default = 15) Height of the plot in inches.}

\item{width}{(int, default = 15) Width of the plot in inches.}

\item{xlim}{(vec, default = c(0,20)) axes limits x-axis}

\item{ylim}{(vec, default = c(0,15))axes limits of y-axis}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{print_plot}{(bool, default = TRUE) Whether to print the plot as an output.}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}
}
\value{

}
\description{
Plot a stacked bar plot of the degree distributions.
}
