% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_degree_regions}
\alias{plot_degree_regions}
\title{Plot the degree distributions across regions}
\usage{
plot_degree_regions(
  e,
  channels = c("cfos", "eyfp"),
  colors = c("red", "green"),
  network = "AD",
  title = "",
  height = 10,
  width = 20,
  ylim = c(0, 15),
  sort_super_region = FALSE,
  region_label_angle = 60,
  label_text_size = 12,
  filter_isolates = TRUE,
  image_ext = ".png",
  print_plot = TRUE,
  save_plot = TRUE,
  theme.bar = NULL
)
}
\arguments{
\item{e}{experiment object}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) Channels to plot}

\item{colors}{(str, default = ) String vector of hexadecimal color codes corresponding to to each channel plotted.}

\item{network}{(str, default = "AD") Which network to plot the degree distribution across regions}

\item{title}{(str, default = "") Plot title.}

\item{height}{(int, default = 15) Height of the plot in inches.}

\item{width}{(int, default = 20) Width of the plot in inches.}

\item{ylim}{(vec, default = c(0,15))axes limits of y-axis}

\item{sort_super_region}{(bool, default = FALSE) Whether to divide into subfacets based on which parent region}

\item{region_label_angle}{(int, default = 60) Angle of region labels.}

\item{label_text_size}{(int, default = 12) Font size of region labels.}

\item{filter_isolates}{(default = TRUE) Avoid plotting isolated nodes (zero value)}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{print_plot}{(bool, default = TRUE) Whether to print the plot as an output.s}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}

\item{theme.bar}{User option to use their own ggplot theme
the experiment object output folder.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Bar plot of degree per region in descending magnitude
}
\examples{
\dontrun{
p <- plot_degree_regions(e, colors = c("#660000", "#FF0000"), channels = "cfos", region_label_angle = 60,
ylim = c(0, 15), image_ext = ".png")
}
}
