% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_heatmaps}
\alias{plot_heatmaps}
\title{Plot correlation heatmaps}
\usage{
plot_heatmaps(
  e,
  by = c("group", "sex"),
  channel = c("cfos", "eyfp", "colabel"),
  colors = c("red", "green", "blue")
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
Plot correlation heatmaps
}
