% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_mean_clust_coeff}
\alias{plot_mean_clust_coeff}
\title{Plot mean clustering coefficient}
\usage{
plot_mean_clust_coeff(
  e,
  color_palettes = c("reds", "greens"),
  colors_manual = NULL,
  channels = c("cfos", "eyfp"),
  labels = c(AD = "AD_label", control = "control_label"),
  title = "my_title",
  height = 10,
  width = 10,
  label_angle = 60,
  rev_x_scale = FALSE,
  ylim = c(0, 0.7),
  theme.gg = NULL,
  image_ext = ".png",
  print_plot = FALSE,
  save_plot = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{color_palettes}{(str, default = c("reds", "greens")) Color palettes from \link[grDevices:palettes]{grDevices::hcl.colors} that are used to for plotting networks characteristics for each channel, respectively.}

\item{colors_manual}{(str, default = NULL ) Manually choose the hexadecimal color codes to create a custom color palette, e.g. colors_manual = c("#660000", "#FF0000", "#FF6666").
Warning: this will be applied to all channels. It's recommended to set the channels parameter to a single channel if this parameter is used.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) Channels to plot}

\item{labels}{(e.g. labels = c(network1_name = "network 1 label", network2_name = "network 2 label)) The legend labels to correspond with your network names.}

\item{title}{(str, default = "my_title) plot title}

\item{height}{(int, default = 10) Height of the plot in inches.}

\item{width}{(int, default = 10) Width of the plot in inches.}

\item{label_angle}{(int, default = 60)}

\item{rev_x_scale}{(bool, default = FALSE) Reveres the scale of the categorical variables
the experiment object output folder.}

\item{ylim}{(vec, default = c(0,10)) Axes limits of y-axis}

\item{theme.gg}{(default = NULL) Option to use custom ggplot2 theme if the user wants}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{print_plot}{(bool, default = FALSE) Whether to print the plot as an output.s}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Plot the mean clustering coefficients of the networks in a barplot. Error bars are plotted as SEM.
}
\examples{
\dontrun{
p <- plot_mean_clust_coeff(e, colors_manual = c("#660000", "#FF0000"), channels = "cfos",
labels = c("AD" = "AD_label", "control" = "control_label"), title = "my title", ylim = c(0, 0.7), image_ext = ".png")
}
}
