% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_networks}
\alias{plot_networks}
\title{Plot the networks stored in an experiment object}
\usage{
plot_networks(
  e,
  network_name = "AD",
  title = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  height = 15,
  width = 15,
  image_ext = ".png",
  print_plot = TRUE,
  save_plot = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{network_name}{(str, default = "AD")}

\item{save_plot}{}
}
\value{

}
\description{
Plot the networks stored in an experiment object
}
