% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_normalized_counts}
\alias{plot_normalized_counts}
\title{Plot normalized cell counts}
\usage{
plot_normalized_counts(
  e,
  channels = c("cfos", "eyfp", "colabel"),
  title = NULL,
  colors = c("#FFFFFF", "lightblue"),
  height = 7,
  width = 20,
  print_plot = TRUE,
  save_plot = TRUE,
  image_ext = ".png"
)
}
\arguments{
\item{e}{experiment object}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel"))}

\item{colors}{(str, default = c()) Hexadecimal codes corresponding to the channels (respectively) to plot.}

\item{height}{height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{print_plot}{(bool, default = TRUE) Whether to display the plot (in addition to saving the plot)}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of
the experiment object output folder.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Plot the cell counts normalized by volume for a given channel
}
