% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_percent_colabel}
\alias{plot_percent_colabel}
\title{This function allows for plotting of colabelled cells over either the "cfos" or "eyfp" channels.}
\usage{
plot_percent_colabel(
  e,
  colabel_channel = "colabel",
  channel = "eyfp",
  rois = c("AAA", "dDG", "HY"),
  color_mapping = "sex",
  colors = c("#952899", "#358a9c"),
  pattern_mapping = NULL,
  patterns = c("gray100", "hs_fdiagonal",
    "hs_horizontal\\n                                              ", "gray90",
    "hs_vertical"),
  error_bar = "sem",
  ylim = c(0, 100),
  plot_individual = TRUE,
  height = 8,
  width = 8,
  print_plot = FALSE,
  save_plot = TRUE,
  image_ext = ".png"
)
}
\arguments{
\item{e}{experiment object}

\item{colabel_channel}{(str, default = "colabel") The channel used as the numerator in fraction counts.}

\item{channel}{(str, default = "eyfp") The channel used as denominator in fraction counts.}

\item{rois}{character vector of region acronyms, e.g. c("AAA", "DG)}

\item{color_mapping}{(str, default = "sex") The name of the categorical variable (e.g., "sex", "age", etc.) to map to the color aesthetic of the bar plot.}

\item{colors}{(str) character vector of the color values desired for the groups.}

\item{pattern_mapping}{(str, default = "sex") The name of the categorical variable (e.g., "sex", "age", etc.) to map to the pattern aesthetic of the bar plot.}

\item{patterns}{(str, default = c("gray100", 'hs_fdiagonal', "hs_horizontal", "gray90", "hs_vertical"), Pattern types to define subgroups.}

\item{error_bar}{(str, c("sd", "sem)) options for which type of error bar to display, standard deviation or standard error of the mean.}

\item{ylim}{(default = c(0,100)) The range of the y-axis}

\item{plot_individual}{(boo) Whether or not to plot multiple}

\item{height}{(default = 8) height of graphics devices in inches}

\item{width}{(default = 8) height of graphics device in inches}

\item{print_plot}{(bool, default = FALSE) whether or not to print the plot for just for display}

\item{save_plot}{(bool, default = TRUE) whether or not to save the plor}

\item{image_ext}{(default = ".png") extension determining the image type to save as}
}
\value{
p Plot handle to the figure
}
\description{
Allows for specification of specific brain regions to plot. Two different mouse attributes can be used as categorical variables to map to either the color or
pattern aesthetics of the bar plot, e.g. sex and experimental group.
The color aesthetic takes precedence over the pattern aesthetic so if you only want to use one mouse attribute, for plotting
set it to the \code{color_mapping} parameter and set the \code{pattern_mapping} parameter to NULL.
}
\examples{
\dontrun{
 plot_percentage_colabel
}
}
