% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.correlation_list}
\alias{print.correlation_list}
\title{Print attributes of correlation_list object}
\usage{
\method{print}{correlation_list}(x, ...)
}
\arguments{
\item{x}{an object used to select a method. A correlation list}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print attributes of correlation_list object
}
