% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.experiment}
\alias{print.experiment}
\title{Print attributes of experiment object}
\usage{
\method{print}{experiment}(e)
}
\arguments{
\item{e}{experiment object}
}
\description{
Print attributes of experiment object
}
