% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mouse}
\alias{print.mouse}
\title{Print attributes of mouse object}
\usage{
\method{print}{mouse}(x, ...)
}
\arguments{
\item{x}{mouse object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print attributes of mouse object
}
\examples{
m <- mouse()
print(m)
}
