% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.slice}
\alias{print.slice}
\title{Print attributes of slice object}
\usage{
\method{print}{slice}(x, ...)
}
\arguments{
\item{x}{slice object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print attributes of slice object
}
\examples{
s <- slice()
print(s)
}
