% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{print.wb_experiment}
\alias{print.wb_experiment}
\title{Print attributes of experiment object}
\usage{
\method{print}{wb_experiment}(e)
}
\arguments{
\item{e}{experiment object}
}
\description{
Print attributes of experiment object
}
