% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{register.mouse}
\alias{register.mouse}
\title{Register a slice in a mouse object}
\usage{
\method{register}{mouse}(
  m,
  slice_ID = "1_10",
  hemisphere = NULL,
  filter = NULL,
  replace = FALSE,
  ...
)
}
\arguments{
\item{m}{mouse object}

\item{slice_ID}{ID of slice (str)}

\item{hemisphere}{'left', 'right' (str) or NULL if both hemispheres are included}

\item{filter}{(list)}

\item{replace}{(bool, default = FALSE) Replace a registration already contained in a mouse object by resetting to NULL value before registration improvement loop.}

\item{...}{additional parameters to pass to the SMART::registration2() function, besides 'input', 'coordinate', 'filter' & 'correspondance'}
}
\value{
m  mouse object
}
\description{
Register a slice in a mouse object
}
\examples{
TODO: add popup window option to better visualize the internal structures of the images
}
