% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R, R/utils.R
\name{register}
\alias{register}
\alias{register.slice}
\alias{register.mouse}
\title{Register (generic function)}
\usage{
register(x, ...)

\method{register}{slice}(s, filter = NULL, ...)

\method{register}{mouse}(
  m,
  slice_ID = NA,
  hemisphere = NULL,
  filter = NULL,
  replace = FALSE,
  ...
)

register(x, ...)

\method{register}{slice}(s, filter = NULL, ...)

\method{register}{mouse}(
  m,
  slice_ID = NA,
  hemisphere = NULL,
  filter = NULL,
  replace = FALSE,
  ...
)
}
\arguments{
\item{x}{a mouse or slice object}

\item{...}{additional parameters to pass to the SMART::registration2() function, besides 'input', 'coordinate', 'filter' & 'correspondance'}

\item{s}{slice object}

\item{filter}{(list) Wholebrain filter with parameters.}

\item{m}{mouse object}

\item{slice_ID}{(str)  ID of slice}

\item{hemisphere}{(str, default = NULL) 'left', 'right' or NULL if both hemispheres are included}

\item{replace}{(bool, default = FALSE) Replace a registration already contained in a mouse object by resetting to NULL value before registration improvement loop.}
}
\value{
a mouse or slice object

s slice object

m  mouse object

s slice object

m  mouse object
}
\description{
Register (generic function)

Register a slice in a slice object

Register a slice in a mouse object. If a slice has been previously registered, the default behavior is to continue modifying the
previous registration. Use the \code{replace} parameter to change this behavior.

Register (generic function)

Register a slice in a slice object

Register a slice in a mouse object. If a slice has been previously registered, the default behavior is to continue modifying the
previous registration. Use the \code{replace} parameter to change this behavior.
}
\examples{
s <- register(s)
m <- register(m, slice_ID = '1_10', hemisphere = "left", filter = my_filter)
s <- register(s)
m <- register(m, slice_ID = '1_10', hemisphere = "left", filter = my_filter)
}
