% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{register.slice}
\alias{register.slice}
\title{Register a slice in a slice object}
\usage{
\method{register}{slice}(s, filter = NULL, ...)
}
\arguments{
\item{s}{slice object}

\item{filter}{(list)}

\item{...}{additional parameters to pass to the SMART::registration2() function, besides 'input', 'coordinate', 'filter' & 'correspondance'}
}
\value{
s slice object
}
\description{
Register a slice in a slice object
}
