% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{reset_mouse_root}
\alias{reset_mouse_root}
\title{Reset the root path for the folder containing the registration and segmentation data.}
\usage{
reset_mouse_root(m, input_path = NULL, print = TRUE)
}
\arguments{
\item{m}{mouse object}

\item{input_path}{(default = NULL) Reset the root directory of the mouse object.}

\item{print}{(bool, default = TRUE) Print the changes in the console.}
}
\value{
m a mouse object
}
\description{
This function takes a mouse object and also a \code{input_path} as the root folder for that mouse.
It then adjusts all the paths for the registration and segmentation data read to be relative to the root folder.
This function is especially useful if you have changed the computers you are analyzing and drive mappings may be different.
}
\examples{
m <- reset_mouse_root(m, input_path = "C:/Users/Documents/Mice/mouse_1/", print = TRUE)
}
