% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{rois_intersect_region_list}
\alias{rois_intersect_region_list}
\title{Get  a list of intersecting regions to a list of common regions}
\usage{
rois_intersect_region_list(common_reg, rois)
}
\arguments{
\item{common_reg}{A comprehensive list of all regions (and all existing subregions) that the brain area in the \code{rois} list will be compared against.}

\item{rois}{A list of rois whose regions and subregions will be compared the the \code{common_reg} list}
}
\value{
common_reg A list of the rois (or any of its subregions) that intersected with the common_reg list.
}
\description{
Get  a list of intersecting regions to a list of common regions
}
