% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simplify.regions}
\alias{simplify.regions}
\title{Modification of Marcos' combine regions function}
\usage{
simplify.regions(
  normalized_counts,
  keywords = c("layer", "part", "stratum", "division")
)
}
\arguments{
\item{normalized_counts}{list with length = No. channels., each channel element is df of normalized counts}

\item{keywords}{a vector of keywords with which to simplify the region names}
}
\value{
a list with length = No. channels. Each channel df has simplified parent acronyms and names is summed from original df based on them.
}
\description{
Modification of Marcos' combine regions function
}
