% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{simplify_by_keywords}
\alias{simplify_by_keywords}
\title{Simplify dataframe by keywords.}
\usage{
simplify_by_keywords(
  df,
  keywords = c("layer", "part", "stratum", "division", "leaflet", "Subgeniculate",
    "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch", "Precommissural")
)
}
\arguments{
\item{df}{(tibble) Must contain columns "acronym" and "name"}

\item{keywords}{(vec, default = c("layer","part","stratum","division")) a list of keywords to simplify based on region name.}
}
\value{
df
}
\description{
Simplify dataframe by keywords.
}
