% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{simplify_cell_count}
\alias{simplify_cell_count}
\title{Simplify the combined cell count table}
\usage{
simplify_cell_count(
  e,
  ontology = "allen",
  simplify_keywords = c("layer", "part", "stratum", "division", "leaflet",
    "Subgeniculate", "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch",
    "Precommissural"),
  dont_fold = c("Dorsal part of the lateral geniculate complex",
    "Ventral posterolateral nucleus of the thalamus, parvicellular part",
    "Ventral posteromedial nucleus of the thalamus, parvicellular part",
    "Ventral posterolateral nucleus of the thalamus, parvicellular part",
    "Ventral posteromedial nucleus of the thalamus, parvicellular part",
    "Substantia nigra")
)
}
\arguments{
\item{e}{experiment object}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}

\item{simplify_keywords}{(str vec, default =  c("layer","part","stratum","division", "leaflet", "Subgeniculate", "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch", "Precommissural")).
Keywords to search through region names and simplify to parent structure. This means the parent structure is also excluded if the list of excluded right and left}

\item{dont_fold}{(vec) vector of regions to not fold in.
regions can be further}
}
\value{
e experiment object with simplified keywords
}
\description{
This function is designed to offer flexible simplification of mapped cells counts. This can be applied after running
\code{\link[=combine_cell_counts]{combine_cell_counts()}}. However, if mapping is being conducted using the SMARTTR package, we recommend simplifying mapped counts earlier, at the level of mouse objects using \code{\link[=normalize_cell_counts]{normalize_cell_counts()}}
because the options offered for simplification are more flexible.
The benefit of this function is that it can operate on experiment objects with externally imported combined cell counts tables that are formatted for compatibility. This allows
for simplification using other ontologies. See the available atlas options under the \code{ontology} parameter.
}
\examples{
\dontrun{
e <- simplify_cell_count(e)
}
}
