% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{slice}
\alias{slice}
\title{Create a slice object}
\usage{
slice_example <- slice() # initializes a slice object
}
\value{
A slice, a colloquial term for an object of class 'slice'. A 'slice' object
is also a list, with class list.
}
\description{
slice() constructs an S3 object of class 'slice'. A slice object
consists of a list of lists storing information about registration, segmentation, raw mapped data and cleaned mapped data.
The object attributes are also stored as a list.
\subsection{Attributes is a list with named initialized values listed below:}{
\enumerate{
\item slice_ID = NA,
\item coordinate = -1,
\item atlas_plate = NA,
\item bin = 1,
\item z_width = 24,            #Measured in um
\item hemisphere = NULL,
\item channels = c('eyfp', 'cfos', 'colabel'),
\item registration_path = 'set registration image path',
\item segmentation_path = 'set segmentation image path',    # Segmentation path may not be used
\item regions_excluded = c("layer 1","PIR1","TR1","PAA1","NLOT1","OT1","MOBgl","OV","VLPO","SO",
"BA","TU","MEAav","ME","TMv","PVp","SUMl","SCzo","fiber tracts","VS")
}
}
}
