% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTR.R
\docType{package}
\name{SMARTR}
\alias{SMARTR}
\title{SMARTR: Wholebrain analysis package in development for the Denny Lab}
\description{
The base of this pipeline is an R package in development with the working title, \code{SMARTR}, a self-referential play on
a previous package developed as an extension to \code{wholebrain} called \code{SMART}.  This package allows for the
user-friendly pre-processing of segmentation data generated from ImageJ to a be compatible with the \code{\link{wholebrain}} package to generate
region-based cell counts that are normalized by volume.
It will also provides tools for data analysis based on experimental groupings.
}
\details{
The package currently allows for easy implementation of the following:
\enumerate{
\item Setting up the pipeline by specifying experimentparameters, and save directories.
\item The interactive registration process.
\item Importing raw segmentation data from .txt files generated from ImageJ for multiple channels.
\item Optionally creating a filter for the 'cfos' and 'eyfp' channels to clean segmented counts.
\item Creating a segmentation object that is compatible with \code{wholebrain} functions.
\item Forward warping and mapping the data onto the standardized mouse atlas.
\item Cleaning the mapped data in all the following ways: + Removing cells that map outside the boundaries of the atlas.
\itemize{
\item Omitting regions by a default list of regions to omit.
\item Omitting regions by user specified region acronyms.
\item Removing Layer 1 cells
\item Removing cells from a contralateral hemisphere per slice if the registrations are divided by right and left hemispheres.
}
\item Obtaining cell counts normalized by region volume (per mm^2^) and region areas (per mm^2^).
}
}
