% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{split_hipp_DV}
\alias{split_hipp_DV}
\title{Split the hippocampal dataset to dorsal and ventral regions}
\usage{
split_hipp_DV(
  m,
  AP_coord = -2.7,
  combine_hemispheres = TRUE,
  merge = TRUE,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division"),
  rois = c("DG", "CA1", "CA2", "CA3")
)
}
\arguments{
\item{m}{mouse object}

\item{AP_coord}{(double) The AP coordinate which separates dorsal hippocampus from ventral hippocampus. Counts anterior to this are considered dorsal, counts posterior to
this are considered ventral.}

\item{combine_hemispheres}{(bool, default = TRUE) Combine normalized cell counts from both hemispheres}

\item{merge}{(bool, not functional) Whether to include the division of dHipp and vHipp into the normalized counts dataframe of the mouse or to store as a separate element for isolated analysis.
Merging will replace the hippocampal counts currently in the normalized counts table. Not merging will create a separate data element in the mouse. TODO: Currently not functional.}

\item{simplify_regions}{(bool, default = TRUE ) simplify the normalized region counts based on keywords in the internal function, \code{simplify_keywords}}

\item{simplify_keywords}{(str vec, default =  c("layer","part","stratum","division")). Keywords to search through region names and simplify to parent structure}

\item{rois}{(vec) List of the region acronyms to include in the hippocampus}
}
\value{
m mouse object with DV cell counts from the hippocampus included
}
\description{
Split the hippocampal dataset to dorsal and ventral regions
}
