% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{summarise_networks}
\alias{summarise_networks}
\title{Summarize multiple networks}
\usage{
summarise_networks(
  e,
  network_names = c("AD", "control"),
  channels = c("cfos", "eyfp", "colabel"),
  save_stats = TRUE,
  save_degree_distribution = TRUE,
  save_betweenness_distribution = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{network_names}{(str, default = NULL) The names of the networks to generate summary tables for.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) The channels to process.}

\item{save_stats}{(bool, default = TRUE) Save the summary stats as a csv file in the output folder}

\item{save_degree_distribution}{(bool, default = TRUE) Save the network degree distributions (frequencies of each degree) across each comparison group as a csv file.}

\item{save_betweenness_distribution}{(bool, default = TRUE) Save the betweenness distribution.}
}
\value{
e experiment object
}
\description{
Create summary dataframes of all networks and create network statistics for multiple networks
}
\examples{
e <- get_network_statistics(e, network_names = c("AD", "control"),
channels = c("cfos", "eyfp", "colabel"),
save_stats = TRUE, save_degree_distribution = TRUE)
}
