% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{summarize_null_networks}
\alias{summarize_null_networks}
\title{Summarize the parameters of the rewired null networks generated by \code{\link[=rewire_network]{rewire_network()}}}
\usage{
summarize_null_networks(
  null_nodes_list,
  network_names = NULL,
  channel = "cfos"
)
}
\arguments{
\item{null_nodes_list}{a list of output summary tables (1 per network) of rewired network properties of all nodes from \code{\link[=rewire_network]{rewire_network()}}.}

\item{network_names}{(str vec) Name of the networks that were rewired (in the same order as the list). \code{null_nodes_list} and \code{network_names} must have
the same length.}

\item{channel}{(str)   channel to process}
}
\value{
a list of length 2. The first element is named \code{global_summary} and contains a table of global summary statistics.
The second element is named \code{node_summary}, and contained per node statistics averaged from multiple null networks.
}
\description{
Summarize the parameters of the rewired null networks generated by \code{\link[=rewire_network]{rewire_network()}}
}
\examples{
\dontrun{
rewire_summary <- rewire_network(e, "Context", channels = "eyfp", return_graphs = FALSE)
summarized_null_networks <- summarize_null_networks(rewire_summary, network_names = "Context", channel = "eyfp")
}
}
