% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_slice}
\alias{add_slice}
\title{Add slice to a mouse object}
\usage{
add_slice(m, s, replace = FALSE)
}
\arguments{
\item{m}{mouse object}

\item{s}{slice object}

\item{replace}{(bool, default = FALSE) Replace a slice already contained in a mouse object.}
}
\description{
Add slice to a mouse object
}
\examples{
m <- mouse()
s <- slice()
m <- add_slice(m, s, replace = FALSE)
m <- add_slice(m, s, replace = TRUE)
}
