% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R, R/utils.R
\name{combine_cell_counts}
\alias{combine_cell_counts}
\title{Combine cell counts across all mice in an experiment into a single dataframe.}
\usage{
combine_cell_counts(e, by)

combine_cell_counts(e, by)
}
\arguments{
\item{e}{experiment object}

\item{by}{(str) names of the experiment attributes (categorical variables) that will be used to create analysis subgroups.}
}
\description{
This function also stores the mouse attribute names (not experiment attributes) as columns that will be used as categorical variables to make analysis subgroups.
The values of these attributes (\code{group}, \code{drug}, \code{age}) will automatically be converted to a string values for consistency.

This function also stores the mouse attribute names (not experiment attributes) as columns that will be used as categorical variables to make analysis subgroups.
The values of these attributes (\code{group}, \code{drug}, \code{age}) will automatically be converted to a string values for consistency.
}
\examples{
e <- combine_cell_counts(e, by = c('groups', 'sex'))
e <- combine_cell_counts(e, by = c('groups', 'sex'))
}
