% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{correlation_diff_permutation}
\alias{correlation_diff_permutation}
\title{This function performs a permutation analysis to compare the region pairwise correlation coefficients between two different analysis groups.}
\usage{
correlation_diff_permutation(
  e,
  correlation_list_name_1,
  correlation_list_name_2,
  channels = c("cfos", "eyfp", "colabel"),
  n_shuffle = 1000,
  method = "pearson",
  seed = 5,
  p_adjust_method = "none",
  alpha = 0.05,
  ...
)
}
\arguments{
\item{e}{experiment object}

\item{correlation_list_name_1}{(str) The name of the correlation list object used as the first group for comparison.}

\item{correlation_list_name_2}{(str) The name of the correlation list object used as the second group for comparison.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) The channels to process.}

\item{n_shuffle}{(int, default = 1000) The number of permutation shuffles.}

\item{method}{(str, default = "pearson", options = c("pearson", "spearman")) Specifies the type of correlations to compute.
Spearman correlations are the Pearson linear correlations computed on the ranks of non-missing elements, using midranks for ties. See also \code{\link[Hmisc:rcorr]{Hmisc::rcorr()}}}

\item{seed}{(int, default = 5) Random seed for future replication.}

\item{p_adjust_method}{(bool or str, default = "none")
Apply the named method to control for the inflated false discovery rate or FWER. Set to FALSE or "none"
to keep "raw" p values. See also \code{\link[stats:p.adjust]{stats::p.adjust()}} for the correction options.}

\item{alpha}{(float, default = 0.05) The alpha cutoff for significance between region pairwise correlation differences}

\item{...}{additional parameters to pass to \code{permute_corr_diff_distrib()}}
}
\value{
e experiment object. The experiment object now has a list called \code{permutation_p_matrix} stored in it. Elements of this \code{permutation_p_matrix} list are
the outputs of different permutation comparison analyses. These elements are named by the groups that were compared.
}
\description{
The data from two different analysis groups are compared by specifying the
\code{correlation_list_name_1} and \code{correlation_list_name_2} parameters. Note that both of these analysis groups must have the same
number of channels to compare. The functions \code{get_correlations()} needs to have been run for each of these analysis groups prior to
running this function. The test statistics used is the pearson values of those in correlation_list_name_2 subtracted from corresponding Pearson values in correlation_list_name_1.
}
\examples{
\dontrun{
e <- correlation_diff_permutation(sundowning,
                                            correlation_list_name_1 = "female_AD",
                                            correlation_list_name_2 = "female_control",
                                            channels = c("cfos", "eyfp", "colabel"),
                                            n_shuffles = 1000,
                                            seed = 5,
                                            p_adjust_method = "none"
                                            alpha = 0.001
                                            )
}

}
\seealso{
\code{\link[=get_correlations]{get_correlations()}}
}
