% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{create_networks}
\alias{create_networks}
\title{Create \%>\% graph objects for plotting different analysis subgroups.}
\usage{
create_networks(
  e,
  correlation_list_name,
  channels = c("cfos", "eyfp", "colabel"),
  proportional_thresh = NULL,
  alpha = 0.05,
  pearson_thresh = 0.8,
  ontology = "allen",
  anatomical.order = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB", "HB",
    "CB"),
  filter_isolates = FALSE
)
}
\arguments{
\item{e}{experiment object}

\item{correlation_list_name}{(str) Name of the correlation list object used to generate the networks.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) The channels to process.}

\item{proportional_thresh}{(float, default = NULL) Takes precedent over the \code{alph}a and the \code{pearson_thresh} parameters. Input the desired edge proportion (i.e., edge density) as your desired sparsity constraint.}

\item{alpha}{(float, default = 0.05) The significance threshold for including brain regions in the network. if NULL or NA,
this threshold is not applied.}

\item{pearson_thresh}{(float, default = 0.8) The pearson correlation coefficient threshold to apply for filtering out}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}

\item{anatomical.order}{(vec, c("Isocortex","OLF","HPF","CTXsp","CNU","TH","HY","MB","HB","CB")) The default super region acronym list that groups all subregions in the dataset.}

\item{filter_isolates}{(logical, default = FALSE) Whether to filter out the number of isolated (zero degree) nodes from the network. Default is to retain them.}
}
\value{
e experiment object. This object now has a new added element called \code{networks.} This is a list storing a
graph object per channel for each network analysis run.
The name of each network (\code{network_name}) is the same as the \code{correlation_list_name}
used to generate the network. This \code{network_name} is fed as a parameter into the
\code{\link[=plot_network]{plot_network()}} function.
}
\description{
Create \%>\% graph objects for plotting different analysis subgroups.
}
\examples{
e sundowning <- create_networks(sundowning, correlation_list_name = "female_control", alpha = 0.05)
}
\seealso{
\code{\link[=plot_networks]{plot_networks()}}
}
