% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{exclude_anatomy}
\alias{exclude_anatomy}
\alias{exclude_anatomy.slice}
\alias{exclude_anatomy.mouse}
\title{exclude_anatomy (generic function)}
\usage{
exclude_anatomy(x, ...)

\method{exclude_anatomy}{slice}(
  s,
  channels = c("cfos", "eyfp", "colabel"),
  clean = TRUE,
  exclude_right_regions = NULL,
  exclude_left_regions = NULL,
  exclude_hemisphere = TRUE,
  exclude_layer_1 = TRUE,
  plot_filtered = TRUE
)

\method{exclude_anatomy}{mouse}(
  m,
  slice_ID = NA,
  hemisphere = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  clean = TRUE,
  exclude_right_regions = NULL,
  exclude_left_regions = NULL,
  exclude_hemisphere = FALSE,
  exclude_layer_1 = TRUE,
  plot_filtered = TRUE
)
}
\arguments{
\item{...}{}

\item{s}{slice object}

\item{channels}{(vec) Channels to process.}

\item{clean}{(bool, default = TRUE ). Remove cells that don't map to any regions.}

\item{exclude_right_regions}{(str vector, default = NULL); acronyms of regions you want to exclude from right hemi,in addition to regions that will by default be excluded in the slice attribute 'right_regions_excluded'}

\item{exclude_left_regions}{(str vector, default = NULL); acronyms of regions you want to exclude from left hemi, in addition to regions that will by default be excluded in the slice attribute 'left_regions_excluded'}

\item{exclude_hemisphere}{(bool, default = TRUE); excludes the contralateral hemisphere from one indicated in slice attribute}

\item{exclude_layer_1}{(bool, default = TRUE); excludes all counts from layer 1 (TEMPORARY, may not be hardcoded in later)}

\item{plot_filtered}{(bool, default = TRUE) pop up window to check the excluded anatomy.}

\item{m}{mouse object}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str) 'left', 'right' or NULL (both)}
}
\value{


m mouse object
}
\description{
Method for excluding user specified regions, layer 1, and contralateral hemisphere  per slice.
This function automatically excludes the default regions included in the attribute "regions_excluded" in each slice
IN ADDITION to the regions added to the 'exclude_regions' parameter.
Regions added to the 'exclude_regions' parameter will then be updated in the slice attribute to keep track of what was excluded.

Method for excluding cell counts in specified regions, layer 1, out-of-bounds cells counts, and hemispheres per mouse.
This function automatically excludes the default regions included in the attribute "regions_excluded" in each slice
IN ADDITION to the regions added to the 'exclude_regions' parameter.
Regions added to the 'exclude_regions' parameter will then be updated in the slice attribute to keep track of what was excluded.
}
\examples{
s <-  exclude_anatomy(s, channels = c('cfos', 'eyfp', 'colabel'), clean = TRUE, exclude_regions = NULL, exclude_hemisphere = TRUE, exclude_layer_1 = TRUE, plot_filtered = TRUE)
m <- exclude_anatomy(m, slice_ID = "1_10", hemisphere = NULL, channels = c('cfos', 'eyfp', 'colabel'), clean = TRUE,
exclude_regions = NULL,
exclude_hemisphere = TRUE,
exclude_layer_1 = TRUE
}
