% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{exclude_anatomy}
\alias{exclude_anatomy}
\alias{exclude_anatomy.slice}
\alias{exclude_anatomy.mouse}
\title{exclude_anatomy (generic function)}
\usage{
exclude_anatomy(x, ...)

\method{exclude_anatomy}{slice}(
  x,
  channels = NULL,
  clean = TRUE,
  exclude_right_regions = NULL,
  exclude_left_regions = NULL,
  exclude_hemisphere = TRUE,
  exclude_layer_1 = TRUE,
  include_right_regions = NULL,
  include_left_regions = NULL,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division", "leaflet",
    "Subgeniculate", "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch",
    "Precommissural"),
  plot_filtered = TRUE,
  ...
)

\method{exclude_anatomy}{mouse}(
  x,
  slice_ID = NA,
  hemisphere = NULL,
  channels = NULL,
  clean = TRUE,
  exclude_right_regions = NULL,
  exclude_left_regions = NULL,
  exclude_hemisphere = FALSE,
  exclude_layer_1 = TRUE,
  include_right_regions = NULL,
  include_left_regions = NULL,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division", "leaflet",
    "Subgeniculate", "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch",
    "Precommissural"),
  plot_filtered = TRUE,
  ...
)
}
\arguments{
\item{x}{mouse object}

\item{...}{further arguments passed to or from other methods.}

\item{channels}{(str vector, default = NULL) Channels to process. If NULL, defaults to the channels stored in the slice object attributes.}

\item{clean}{(bool, default = TRUE ). Remove cells that don't map to any regions.}

\item{exclude_right_regions}{(str vector, default = NULL); acronyms of regions you want to exclude from right hemi,in addition to regions that will by default be excluded in the slice attribute 'right_regions_excluded'}

\item{exclude_left_regions}{(str vector, default = NULL); acronyms of regions you want to exclude from left hemi, in addition to regions that will by default be excluded in the slice attribute 'left_regions_excluded'}

\item{exclude_hemisphere}{(bool, default = TRUE); excludes the contralateral hemisphere from one indicated in slice attribute}

\item{exclude_layer_1}{(bool, default = TRUE); excludes all counts from layer 1 (TEMPORARY, may not be hardcoded in later)}

\item{include_right_regions}{(str vector, default = NULL) Acronyms of regions to include from the right hemi; if not NULL, takes precedence over \code{exclude_right_regions} & all other regions will be excluded. Typically, this is used for slices with poor quality/lots of tears.}

\item{include_left_regions}{(str vector, default = NULL) Acronyms of regions to include from the light hemi; if not NULL, takes precedence over \code{exclude_left_regions} & all other regions will be excluded.  Typically, this is used for slices with poor quality/lots of tears.}

\item{simplify_regions}{(bool, default = TRUE ) simplify the normalized region counts based on keywords in the internal function, \code{simplify_keywords}}

\item{simplify_keywords}{(str vec, default =  c("layer","part","stratum","division")). Keywords to search through region names and simplify to parent structure. This means the parent structure is also excluded if the list of excluded right and left
regions can be further}

\item{plot_filtered}{(bool, default = TRUE) plot the segmented cells following the filtering process}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str) 'left', 'right' or NULL (both)}
}
\value{
a mouse or slice object

m mouse object
}
\description{
Method for excluding user specified regions, layer 1, and contralateral hemisphere  per slice.
This function automatically excludes the default regions included in the attribute "regions_excluded" in each slice
IN ADDITION to the regions added to the 'exclude_regions' parameter.
Regions added to the 'exclude_regions' parameter will then be updated in the slice attribute to keep track of what was excluded.
Note: Please see the \code{simplify_regions} and \code{simplify_keywords} parameters. By default, if a subregion can be folded into a parent region based on a certain keywords, then this function will automatically
exclude the entire parent region as a conservative exclusion approach. Keep \code{simplify_regions=TRUE} if the final analysis will contain simplified regions.

Method for excluding cell counts in specified regions, layer 1, out-of-bounds cells counts, and hemispheres per mouse.
This function automatically excludes the default regions included in the attribute "regions_excluded" in each slice
IN ADDITION to the regions added to the 'exclude_regions' parameter.
Regions added to the 'exclude_regions' parameter will then be updated in the slice attribute to keep track of what was excluded.

Note: Please see the \code{simplify_regions} and \code{simplify_keywords} parameters. By default, if a subregion can be folded into a parent region based on a certain keywords, then this function will automatically
exclude the entire parent region as a conservative exclusion approach. Keep \code{simplify_regions=TRUE} if the final analysis will contain simplified regions.
}
\examples{
\dontrun{
 s <-  exclude_anatomy(s, channels = c('cfos', 'eyfp', 'colabel'), clean = TRUE, exclude_regions = NULL, exclude_hemisphere = TRUE, exclude_layer_1 = TRUE, plot_filtered = TRUE)
}
\dontrun{
m <- exclude_anatomy(m, slice_ID = "1_10", hemisphere = NULL, channels = c('cfos', 'eyfp', 'colabel'), clean = TRUE,
exclude_regions = NULL,
exclude_hemisphere = TRUE,
exclude_layer_1 = TRUE
}
}
