% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R, R/utils.R
\name{find_segmentation_files}
\alias{find_segmentation_files}
\title{Find segmentation files following the naming conventions of the denny lab given a channel name and a root slice directory}
\usage{
find_segmentation_files(slice_directory, channel)

find_segmentation_files(slice_directory, channel)
}
\arguments{
\item{slice_directory}{path to the slice directory}

\item{channel}{which channel to import files for}
}
\value{
returns a vector of two paths. The first element is the path to the measurement data. The second path is the path to the
quantification data.

returns a vector of two paths. The first element is the path to the measurement data. The second path is the path to the
quantification data.
}
\description{
Find segmentation files following the naming conventions of the denny lab given a channel name and a root slice directory

Find segmentation files following the naming conventions of the denny lab given a channel name and a root slice directory
}
