% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{find_segmentation_files}
\alias{find_segmentation_files}
\title{Find segmentation files following the naming conventions of the denny lab given a channel name and a root slice directory}
\usage{
find_segmentation_files(slice_directory, channel)
}
\value{
returns a vector of two paths. The first element is the path to the measurement data. The second path is the path to the
quantification data.
}
\description{
Find segmentation files following the naming conventions of the denny lab given a channel name and a root slice directory
}
