% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get.colabeled.cells}
\alias{get.colabeled.cells}
\title{Get colabelled cells data table}
\usage{
get.colabeled.cells(
  coloc.table,
  eyfp.counts,
  eyfp.counts.16bit,
  volume = 25,
  overlap = 0.5
)
}
\arguments{
\item{overlap}{}
}
\value{
returns a dataframe of colabelled cell counts
}
\description{
Get colabelled cells data table
}
