% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R, R/utils.R
\name{get.registered.areas.bu}
\alias{get.registered.areas.bu}
\title{Get the registered areas}
\usage{
get.registered.areas.bu(regions, registration, conversion.factor = 1)

get.registered.areas.bu(regions, registration, conversion.factor = 1)
}
\arguments{
\item{conversion.factor}{}
}
\description{
Get the registered areas

Get the registered areas
}
