% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R, R/utils.R
\name{get.registered.areas.td}
\alias{get.registered.areas.td}
\title{Get top down registered areas}
\usage{
get.registered.areas.td(regions, registration, conversion.factor = 1)

get.registered.areas.td(regions, registration, conversion.factor = 1)
}
\description{
Get top down registered areas

Get top down registered areas
}
