% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R, R/utils.R
\name{get.registered.areas.td}
\alias{get.registered.areas.td}
\title{Get top down registered areas}
\usage{
get.registered.areas.td(regions, registration, conversion.factor = 1)

get.registered.areas.td(regions, registration, conversion.factor = 1)
}
\arguments{
\item{regions}{vector of acronyms of regions in the registration}

\item{registration}{wholebrain registration object}

\item{conversion.factor}{pixel-to-micron conversion factor}
}
\value{
tibble dataframe with the columns \code{name}, \code{acronym}, \code{right.hemisphere}, and  \code{area}
}
\description{
Get top down registered areas

Get top down registered areas
}
