% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check_functions.R
\name{get_common_regions}
\alias{get_common_regions}
\title{Get common regions that are found across all the comparison groups.}
\usage{
get_common_regions(mouse_count, by)
}
\arguments{
\item{by}{}
}
\value{
unique acronyms that are the intersection across all groups
}
\description{
If a single comparison group doesn't have any mice that have brain regions
contained in the other groups, no statistical comparison can be made. This region
should be removed.
}
