% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{get_correlations}
\alias{get_correlations}
\title{get_correlations}
\usage{
get_correlations(
  e,
  by = c("sex", "group"),
  values = c("female", "AD"),
  channels = c("cfos", "eyfp", "colabel"),
  method = "pearson",
  ...
)
}
\arguments{
\item{e}{experiment object}

\item{by}{The variable names (columns) to filter the dataframe selection.}

\item{values}{The value of the variables to filter the combined normalized counts df by.}

\item{channels}{The channels to correlate.}

\item{method}{Correlational method to pass to \code{\link[corrr:correlate]{corrr::correlate()}} function.}

\item{, , , }{additional parameters to pass to the \code{\link[=corrr:correlate]{corrr:correlate()}} function aside from method, x, and y.}
}
\value{

}
\description{
get_correlations
}
