% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{get_correlations}
\alias{get_correlations}
\title{get_correlations}
\usage{
get_correlations(
  e,
  by = c("sex", "group"),
  values = c("female", "AD"),
  channels = c("cfos", "eyfp", "colabel"),
  p_adjust_method = "BH",
  alpha = 0.05
)
}
\arguments{
\item{e}{experiment object}

\item{by}{The variable names (columns) to filter the dataframe selection.}

\item{values}{The value of the variables to filter the combined normalized counts df by.}

\item{channels}{The channels to correlate.}

\item{p_adjust_method}{(bool or str, default = "BH") Benjamini-Hochberg method is recommended.
Apply the named method to control for the inflated false discovery rate or FWER. Set to FALSE
to keep "raw" p values.}

\item{alpha}{The alpha level for p adjustment.}
}
\value{
e experiment object. The experiment object now has a named \code{correlation_list} object stored in it.
The name of the correlation object is the concatenation of the variable values separated by a "_".
This name allows for unambiguous identification of different group correlations analysis in the future.
}
\description{
get_correlations
}
