% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{get_correlations}
\alias{get_correlations}
\title{Get regional cross correlations and their p-values in a correlation list object.}
\usage{
get_correlations(
  e,
  by,
  values,
  channels = c("cfos", "eyfp", "colabel"),
  p_adjust_method = "none",
  alpha = 0.05,
  ontology = "allen",
  method = "pearson",
  anatomical.order = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB", "HB",
    "CB"),
  region_order = NULL
)
}
\arguments{
\item{e}{experiment object}

\item{by}{(str) Attribute names to group by, e.g. c("sex", "group")}

\item{values}{(str) The respective values of the attributes entered for the \code{by} parameter to generate a specific analysis group,
e.g.values = c("female", "AD").}

\item{channels}{(str, channels =  c("cfos", "eyfp", "colabel") The channels to process.}

\item{p_adjust_method}{(bool or str, default = "none") This parameter is fed into the p.adjust function. Options: c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY",  "fdr", "none")
Apply the named method to control for the inflated false discovery rate or family wise error rate (FWER). Set to FALSE or "none"
to keep "raw" p values. See also \code{\link[stats:p.adjust]{stats::p.adjust()}} for the correction options.}

\item{alpha}{(num, default = 0.05) The alpha level for significance applied AFTER p-adjustment.}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}

\item{method}{(str, default = "pearson", options = c("pearson", "spearman")) Specifies the type of correlations to compute. Spearman correlations are the Pearson linear correlations computed on the ranks of non-missing elements,}

\item{anatomical.order}{(str, default = c("Isocortex","OLF","HPF","CTXsp","CNU","TH","HY","MB","HB","CB")) The order of the regions to plot in the heatmaps.
using midranks for ties. See also \code{\link[Hmisc:rcorr]{Hmisc::rcorr()}}}

\item{region_order}{(list, default = NULL)  optional list with the first element named "acronym" supplying a vector as region acronyms and the second element named
"order"  supplying an vector of integers determining numerical order, e.g. 1, 1, 2, 2.}
}
\value{
e experiment object. The experiment object now has a named \code{correlation_list} object stored in it.
The name of the correlation object is the concatenation of the variable values separated by a "_".
This name allows for unambiguous identification of different analysis subgroups in the future.
}
\description{
This analysis will get regional cross correlations based on cell counts normalized by region volume.
}
\examples{
\dontrun{
e <- get_correlations(e, by = c("sex", "group"), values = c("female", "AD"),
channels = c("cfos", "eyfp", "colabel"),  p_adjust_method = "BH", alpha = 0.05)
}
}
\seealso{
\code{\link[Hmisc:rcorr]{Hmisc::rcorr()}}
}
