% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{get_percent_colabel}
\alias{get_percent_colabel}
\title{Get the percentage of colabelled cells over either cfos or eyfp channels.}
\usage{
get_percent_colabel(
  e,
  by,
  colabel_channel = "colabel",
  channel = "eyfp",
  save_table = TRUE,
  rois = NULL,
  individual = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{by}{(str) Attribute names to group by, e.g. by = c("group", "sex"). These will generate analysis subgroups that are
averaged together to assess across all rois.}

\item{colabel_channel}{(str, default = "colabel") The channel used as the numerator in fraction counts. The string 'colabel' in the pipeline
refers to colocalized 'eyfp' and 'cfos' channels. For other colocalized channels, import the channel using \code{\link[=import_segmentation_custom]{import_segmentation_custom()}}
or your own customized import channel.}

\item{channel}{(str, default = "eyfp") The channel used as denominator in fraction counts.}

\item{save_table}{(bool, default = TRUE) Whether to save the output table as a csv in the experiment object output folder.}

\item{rois}{(str, default = NULL) Whether to generate colabelled percentages for only specific regions of interest, e.g. rois = c("HY", "DG").
Child regions of specified rois will also be searched for.}

\item{individual}{(bool, default = FALSE) Whether the data should include individual mouse colabelled percentages rather than the average.
If FALSE the colabel percentages are averaged across all analysis subgroups determined by the \code{by} parameter}
}
\value{
e experiment object with colabelled percentage table stored in it.
}
\description{
This analysis will only include common regions that are included in both the colabelled and
cfos or eyfp channels. The colabelled percentage of individual animals will be calculated with the option to export the data.
}
\examples{
e <- get_percent_colabel(e, c("group", "sex", channel = "eyfp"))
}
