% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{get_registered_areas}
\alias{get_registered_areas}
\alias{get_registered_areas.slice}
\alias{get_registered_areas.mouse}
\title{get_registered_areas (generic function)}
\usage{
get_registered_areas(x, ...)

\method{get_registered_areas}{slice}(s)

\method{get_registered_areas}{mouse}(m, slice_ID, hemisphere = NULL, replace = FALSE)
}
\arguments{
\item{...}{}

\item{s}{slice object}

\item{m}{mouse object}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str) 'left', 'right' or NULL (both)}
}
\value{


s slice object with a stored dataframe with columns 'name' (full region name), 'acronym', 'area' (in microns^2^), 'right.hemisphere'

m mouse object
}
\description{
Calculate the registered area (in microns^2^) of all regions contained in a slice.
}
\examples{
s <- get_registered_areas(s)
m <- get_registered_areas(m, slice_ID = "1_10", hemisphere = "left", replace = FALSE)
}
