% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{get_registered_volumes}
\alias{get_registered_volumes}
\alias{get_registered_volumes.slice}
\title{get_registered_areas (generic function)}
\usage{
get_registered_volumes(x, ...)

\method{get_registered_volumes}{slice}(s)
}
\arguments{
\item{...}{}

\item{s}{slice object}
}
\value{


s slice object with a stored dataframe with columns 'name' (full region name), 'acronym', 'area' (in microns^2^), 'volume' (in microns^3^) 'right.hemisphere'
}
\description{
Calculate the registered area (microns^2^) and the regional volumes (microns^3^) of all regions contained in a slice.
}
\examples{
s <- get_registered_areas(s)
}
