% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{import_segmentation_ij}
\alias{import_segmentation_ij}
\alias{import_segmentation_ij.slice}
\alias{import_segmentation_ij.mouse}
\title{import_segmentation (generic function)}
\usage{
import_segmentation_ij(x, ...)

\method{import_segmentation_ij}{slice}(s, mouse_ID = NA, channels = NULL, maxdist = 10)

\method{import_segmentation_ij}{mouse}(
  m,
  slice_ID = NA,
  hemisphere = NULL,
  channels = NULL,
  maxdist = 10,
  replace = FALSE
)
}
\arguments{
\item{x}{a mouse or slice object}

\item{...}{further arguments passed to or from other methods.}

\item{s}{slice object}

\item{mouse_ID}{(str) ID of mouse}

\item{channels}{(str vector, default = NULL) channels to import. If NULL, defaults to the channels stored in the slice object attributes.}

\item{maxdist}{(int, default = 10) maximum tolerability of character differences to match the string names of the importation files}

\item{m}{mouse object}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str)'left', 'right' or NULL}

\item{replace}{(bool, default = FALSE) replace existing raw segmentation data}
}
\value{
a mouse or slice object

s slice object

m mouse object
}
\description{
Method for importing segmentation data for a slice object

Method for importing segmentation data for a mouse object
}
\note{
The designated \code{colabel} channel name in this pipeline will auto import the output of the batch_3D_MultiColocalization.ijm macro provided in the pre-processing pipeline.
If you have a separate method used for detecting colabelled cells, please use a different naming convention for this channel,
e.g. "colabel_PV_cfos", and import using a customized import function such as \code{\link[=import_segmentation_custom]{import_segmentation_custom()}}.
}
\examples{
s <-  import_segmentation(s, mouse_ID = "255") # Defaults to channels stored in slice attributes
s <-  import_segmentation(s, mouse_ID = "255", channels = c("cfos", "eyfp", "colabel")) # Specify channels
m <-  import_segmentation(m, slice_ID = "1_10", channels = c("cfos", "eyfp", "colabel"), replace = FALSE)
}
