% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{make.filter}
\alias{make.filter}
\title{#' Modification of Marcos' combine regions function
#'
#' @param normalized_counts list with length = No. channels., each channel element is df of normalized counts
#' @param keywords a vector of keywords with which to simplify the region names
#'
#' @return a list with length = No. channels. Each channel df has simplified parent acronyms and names is summed from original df based on them.
#'
#' @examples
simplify.regions <- function(normalized_counts, keywords = c("layer","part","stratum","division")) {}
\usage{
make.filter(
  data,
  params = c("Vol..unit.", "Moment1", "Moment2", "Moment3", "Moment4", "Sigma"),
  ranges = list(c(200, 12000), c(3, 50), c(0, 600), c(0, 2000), c(0, 5), c(20, Inf))
)
}
\value{

}
\description{
#' Modification of Marcos' combine regions function
#'
#' @param normalized_counts list with length = No. channels., each channel element is df of normalized counts
#' @param keywords a vector of keywords with which to simplify the region names
#'
#' @return a list with length = No. channels. Each channel df has simplified parent acronyms and names is summed from original df based on them.
#'
#' @examples
simplify.regions <- function(normalized_counts, keywords = c("layer","part","stratum","division")) {
}
\details{

}
\section{Initialize empty list vector to store the simplified counts}{
simplified_counts <- vector(mode = "list", length = length(normalized_counts))
names(simplified_counts) <- names(normalized_counts)
}

