% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{make.filter}
\alias{make.filter}
\title{Make a filter for the cfos or eyfp channel}
\usage{
make.filter(
  data,
  params = c("Vol..unit.", "Moment1", "Moment2", "Moment3", "Moment4", "Sigma"),
  ranges = list(c(200, 12000), c(3, 50), c(0, 600), c(0, 2000), c(0, 5), c(20, Inf))
)
}
\arguments{
\item{ranges}{}
}
\value{

}
\description{
Make a filter for the cfos or eyfp channel
}
