% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{make_segmentation_object}
\alias{make_segmentation_object}
\alias{make_segmentation_object.slice}
\alias{make_segmentation_object.mouse}
\title{make_segmentation_object (generic funciton)}
\usage{
make_segmentation_object(x, ...)

\method{make_segmentation_object}{slice}(
  x,
  mouse_ID = NA,
  channels = NULL,
  use_filter = FALSE,
  ...
)

\method{make_segmentation_object}{mouse}(
  x,
  slice_ID = NA,
  hemisphere = NULL,
  channels = NULL,
  replace = FALSE,
  use_filter = FALSE,
  ...
)
}
\arguments{
\item{x}{a mouse or slice object}

\item{...}{(optional) additional volume and overlap parameters for get.colabeled.cells().}

\item{mouse_ID}{(str) ID of mouse}

\item{channels}{(str vector, default = NULL) Channels to process. If NULL, defaults to the channels stored in the slice object attributes.}

\item{use_filter}{(bool, default = FALSE). Use a filter to create more curated segmentation object from the raw segmentation data.}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str, default = NULL) 'left', 'right' or NULL (both)}

\item{replace}{(bool, default = FALSE) replace existing raw segmentation data}
}
\value{
a mouse or slice object

s slice object
}
\description{
Make a wholebrain compatible segmentation object for a slice in a slice object

Make a wholebrain compatible segmentation object for a slice in a mouse object
}
\note{
If you are processing the colabel channel, the X and Y positions of colabelled cells are the average of the x,y centroid coordinates of the colabelled objects
}
\examples{
\dontrun{
s <- make_segmentation_object(s, mouse_ID = "255",
channels = c("cfos", "eyfp"), use_filter = FALSE)
}
\dontrun{
m <-  make_segmentation_object(m, slice_ID = '1_9', hemisphere = 'left',
channels = c('eyfp', 'cfos', 'colabel'), use_filter = FALSE)
}
}
