% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{make_segmentation_object}
\alias{make_segmentation_object}
\alias{make_segmentation_object.slice}
\alias{make_segmentation_object.mouse}
\title{make_segmentation_object (generic funciton)}
\usage{
make_segmentation_object(x, ...)

\method{make_segmentation_object}{slice}(
  s,
  mouse_ID = NA,
  channels = c("cfos", "eyfp", "colabel"),
  use_filter = FALSE,
  ...
)

\method{make_segmentation_object}{mouse}(
  m,
  slice_ID = NA,
  hemisphere = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  replace = FALSE,
  use_filter = FALSE,
  ...
)
}
\arguments{
\item{...}{additional volume and overlap parameters for get.colabeled.cells().}

\item{s}{slice object}

\item{mouse_ID}{(str) ID of mouse}

\item{channels}{(vec) Channels to process.}

\item{use_filter}{(bool, default = FALSE). Use a filter to create more curated segmentation object from the raw segmentation data.}

\item{m}{mouse object}

\item{slice_ID}{(str) ID of slice}

\item{hemisphere}{(str) 'left', 'right' or NULL (both)}

\item{replace}{(bool, default = FALSE) replace existing raw segmentation data}
}
\value{


s slice object
}
\description{
Make a wholebrain compatible segmentation object for a slice in a slice object

Make a wholebrain compatible segmentation object for a slice in a mouse object
}
\note{
If you are processing the colabel channel, you must always have the raw eyfp segmentation data imported.
}
\examples{
s <- make_segmentation_object(s, mouse_ID = "255", channels = c("cfos", "eyfp"), use_filter = FALSE)
m <-  make_segmentation_object(m, slice_ID = '1_9', hemisphere = 'left', channels = c('eyfp', 'cfos'), use_filter = FALSE)
}
