% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{mouse}
\alias{mouse}
\title{Create a mouse object}
\usage{
mouse_example <- mouse() # initializes a mouse object
}
\value{
A mouse, a colloquial term for an object of class 'mouse'. A 'mouse' object
is also a list, with class list.
}
\description{
mouse() constructs an S3 object of class 'mouse'. A mouse object
consists of a list of slice objects and attributes stored as a list.
\subsection{Attributes are strings with initialized values listed below:}{
\enumerate{
\item mouse_ID = 'set ID'
\item experiment = 'create experiment name'
\item group = 'name experiment group',
\item cohort = 'name mouse cohort'
\item input_path = 'set input path'
\item output_path = 'set output path'
}
}
}
