% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{normalize.registered.areas}
\alias{normalize.registered.areas}
\title{normalize registered areas}
\usage{
normalize.registered.areas(cell_table, total_volumes)
}
\arguments{
\item{cell_table}{cell_tables list from get_cell_table()}

\item{total_volumes}{computed total volumes list}
}
\value{

}
\description{
normalize registered areas
}
