% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{normalize_cell_counts}
\alias{normalize_cell_counts}
\title{Normalize cell counts per mm^2^ or by mm^3^ (if multiplying by the stack size).}
\usage{
normalize_cell_counts(
  m,
  combine_hemispheres = TRUE,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division")
)
}
\arguments{
\item{m}{mouse object}

\item{combine_hemispheres}{(bool, default = TRUE) Combine normalized cell counts from both hemispheres}

\item{simplify_regions}{(bool, default = TRUE ) simplify the normalized region counts based on keywords in the internal function, \code{simplify_keywords}}

\item{simplify_keywords}{(str vec, default =  c("layer","part","stratum","division")). Keywords to search through region names and simplify to parent structure}
}
\value{

}
\description{
Run this function after all the slices that you want to process are finished being added
and you have combined your cell counts with \code{\link[=get_cell_table]{get_cell_table()}}. This functions process all channels
where a cell table was made using the latter function.
}
\examples{
m <- normalize_cell_counts(m, combine_hemispheres = TRUE, simplify_regions = TRUE)
}
