% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{normalize_cell_counts}
\alias{normalize_cell_counts}
\title{Normalize cell counts per mm^^2 or by mm^3^ (if multiplying by the stack size)}
\usage{
normalize_cell_counts(
  m,
  combine_hemispheres = TRUE,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division")
)
}
\arguments{
\item{m}{mouse object}

\item{combine_hemispheres}{bool : Combine normalized cell counts from both hemispheres}

\item{simplify_regions}{bool : simplify the normalized region counts based on keywords in \code{simplify_keywords}}

\item{simplify_keywords}{str vector : c("layer","part","stratum","division"), keywords to search through region names and simplify to parent structure}
}
\value{

}
\description{
Run this function after all the slices that you want to process are finished being added
and you have run the function get_cell_table() for your mouse.
}
\examples{
m <- normalize_cell_counts(m, combine_hemispheres = TRUE, simplify_regions = TRUE)
}
