% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{parallel_coordinate_plot}
\alias{parallel_coordinate_plot}
\title{Create a parallel coordinate plot}
\usage{
parallel_coordinate_plot(
  e,
  permutation_comparison = "AD_vs_control",
  channels = c("cfos", "eyfp", "colabel"),
  colors = c("#be0000", "#00782e", "#f09b08"),
  x_label_group_1 = NULL,
  x_label_group_2 = NULL,
  height = 10,
  width = 10,
  print_plot = TRUE,
  save_plot = TRUE,
  reverse_group_order = FALSE,
  force = 1,
  plt_theme = NULL,
  label_size = 30,
  image_ext = ".png",
  nudge_x = 2:5
)
}
\arguments{
\item{e}{experiment object}

\item{permutation_comparison}{The name of the correlation group comparisons to plot.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) channels to plot}

\item{colors}{(str, default = c("#be0000", "#00782e", "#f09b08")) Hexadecimal codes corresponding to the channels (respectively) to plot.}

\item{x_label_group_1}{(str, NULL) The label for the first group in the permutation analysis. Note: this is to customize the graph labels. It does not reverse the group order.}

\item{x_label_group_2}{(str, NULL) The label for the second group in the permutaiton analysis. Note: this is to customize the graph labels. It does not reverse the group order.}

\item{height}{height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{print_plot}{(bool, default = TRUE) Whether to display the plot (in addition to saving the plot)}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}

\item{reverse_group_order}{(bool, default = TRUE) Reverse the order of the groups on the x-axis.}

\item{force}{(default =1) Force of the text repel between text labels.}

\item{plt_theme}{(default = NULL) Add a \code{\link[ggplot2:theme]{ggplot2::theme()}} to the plot. If NULL, the default is taken.}

\item{label_size}{(default = 30) Default font size for region labels.}

\item{image_ext}{(default = ".png") image extension to save the plot as.}

\item{nudge_x}{(vec, default = 2:5) a vector determining the jitter between labels.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Plot the correlation difference between two comparison groups into a parallel coordinate plot. The function
\code{\link[=correlation_diff_permutation]{correlation_diff_permutation()}} must be run first in order to generate results to plot.
}
\examples{
\dontrun{
p_list <- parallel_coordinate_plot(e, permutation_comparison = "Context_vs_Shock",
 channels = "cfos", colors ="#be0000", x_label_group1 = "Context", x_label_group_2 = "Shock")
}
}
