% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_correlation_heatmaps}
\alias{plot_correlation_heatmaps}
\title{Plot correlation heatmaps}
\usage{
plot_correlation_heatmaps(
  e,
  correlation_list_name,
  colors = c("#be0000", "#00782e", "#f09b08"),
  print_plot = TRUE,
  save_plot = TRUE,
  image_ext = ".png",
  title = NULL,
  height = 10,
  width = 10
)
}
\arguments{
\item{e}{experiment object. Must contain a named correlation_list object generated by \code{\link[=get_correlations]{get_correlations()}}}

\item{correlation_list_name}{(str) The name of the correlation object generated by \code{\link[=get_correlations]{get_correlations()}}}

\item{colors}{(str, default = c("#be0000", "#00782e", "#f09b08")) Hexadecimal code for the colors corresponding to the channels attribute of the correlation_list. Color values can also be
input compatible with ggplot2 plotting functions.}

\item{print_plot}{(bool, default = TRUE) Print the plot as graphics windows.}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{title}{(str, default = NULL) If NULL, the \code{correlation_list_name} will used as the title with underscores removed.}

\item{height}{(int) Height of the plot in inches.}

\item{width}{(int) Width of the plot in inches.}
}
\description{
Plot correlation heatmaps
}
\examples{
plot_correlation_heatmaps(e, correlation_list_name = "female_AD") # No return value
}
\seealso{
\code{\link[=get_correlations]{get_correlations()}}
}
