% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_correlation_heatmaps}
\alias{plot_correlation_heatmaps}
\title{Plot correlation heatmaps}
\usage{
plot_correlation_heatmaps(
  e,
  correlation_list_name,
  channels = c("cfos", "eyfp", "colabel"),
  colors = c("#be0000", "#00782e", "#f09b08"),
  sig_color = "yellow",
  sig_nudge_y = -0.7,
  sig_size = 7,
  print_plot = TRUE,
  save_plot = TRUE,
  image_ext = ".png",
  plot_title = NULL,
  height = 10,
  width = 10,
  theme.hm = NULL
)
}
\arguments{
\item{e}{experiment object. Must contain a named correlation_list object generated by \code{\link[=get_correlations]{get_correlations()}}}

\item{correlation_list_name}{(str) The name of the correlation object generated by \code{\link[=get_correlations]{get_correlations()}}}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) Must exist in the channels attribute of the correlation_list.}

\item{colors}{(str, default = c("#be0000", "#00782e", "#f09b08")) Hexadecimal code for the colors corresponding channels parameter. Color values can also be
input compatible with ggplot2 plotting functions.}

\item{sig_color}{(str, default = "yellow") Color of the significance symbol in R}

\item{sig_nudge_y}{(default = -0.7) Relative amount to nudge the significance symbols in the y direction to center over each square.}

\item{sig_size}{(default = 7) Point size for significance symbol.}

\item{print_plot}{(bool, default = TRUE) Print the plot as graphics windows.}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{plot_title}{(str, default = NULL) If NULL, the \code{correlation_list_name} will used as the title with underscores removed.}

\item{height}{(int) Height of the plot in inches.}

\item{width}{(int) Width of the plot in inches.}

\item{theme.hm}{(default = NULL) Option to use custom ggplot2 theme if the user wants}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Plot correlation heatmaps
}
\examples{
plot_correlation_heatmaps(e, correlation_list_name = "female_AD") # No return value
}
\seealso{
\code{\link[=get_correlations]{get_correlations()}}
}
