% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_degree_regions}
\alias{plot_degree_regions}
\title{Plot the degree distributions across regions}
\usage{
plot_degree_regions(
  e,
  channels = c("cfos", "eyfp"),
  colors = c("red", "green"),
  network = "AD",
  title = "my_title",
  height = 10,
  width = 20,
  ylim = c(0, 15),
  image_ext = ".png",
  print_plot = TRUE,
  save_plot = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) Channels to plot}

\item{colors}{(str, default = ) String vector of hexadecimal color codes corresponding to to each channel plotted.}

\item{network}{(str, default = "AD") Which network to plot the degree distribution across regions}

\item{title}{the experiment object output folder.}

\item{height}{(int, default = 15) Height of the plot in inches.}

\item{width}{(int, default = 20) Width of the plot in inches.}

\item{ylim}{(vec, default = c(0,15))axes limits of y-axis}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{print_plot}{(bool, default = TRUE) Whether to print the plot as an output.s}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}
}
\value{

}
\description{
Bar plot of degree per region in descending magnitude
}
