% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_networks}
\alias{plot_networks}
\title{Plot the networks stored in an experiment object}
\usage{
plot_networks(
  e,
  network_name = "AD",
  title = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  edge_color = "firebrick",
  height = 15,
  width = 15,
  image_ext = ".png",
  print_plot = TRUE,
  save_plot = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{network_name}{(str, default = "AD")}

\item{title}{(str, default = NULL) Title of network plot}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel"))}

\item{edge_color}{(str, default = "firebrick") Color of the network edges.
Can also be a hexadecimal color code written as a string.}

\item{height}{Height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{print_plot}{(bool, default = TRUE) Whether to print the plot as an output.
the experiment object output folder.}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}
}
\description{
Plot the networks stored in an experiment object
}
