% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_networks}
\alias{plot_networks}
\title{Plot the networks stored in an experiment object}
\usage{
plot_networks(
  e,
  network_name = "AD",
  title = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  edge_color = "firebrick",
  height = 15,
  width = 15,
  degree_scale_limit = c(1, 10),
  correlation_edge_width_limit = c(0.8, 1),
  image_ext = ".png",
  network_radius = 300,
  print_plot = TRUE,
  graph_theme = NULL,
  label_size = 5,
  label_offset = 0.15,
  save_plot = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{network_name}{(str, default = "AD")}

\item{title}{(str, default = NULL) Title of network plot}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel"))}

\item{edge_color}{(str, default = "firebrick") Color of the network edges.}

\item{height}{Height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{degree_scale_limit}{(vec, default = c(1,10)) Scale limit for degree size}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{print_plot}{(bool, default = TRUE) Whether to print the plot as an output.
the experiment object output folder.}

\item{graph_theme}{(default = NULL) Add a \code{\link[ggraph:theme]{ggraph::theme()}} to the network graph. If NULL, the default is taken.}

\item{label_size}{(default = 5) Default font size for network region labels.}

\item{label_offset}{(default = 0.15) Distance of label from nodes.
Can also be a hexadecimal color code written as a string.}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Plot the networks stored in an experiment object
}
