% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_networks}
\alias{plot_networks}
\title{Plot the networks stored in an experiment object}
\usage{
plot_networks(
  e,
  network_name = "AD",
  title = NULL,
  channels = c("cfos", "eyfp", "colabel"),
  edge_color = "firebrick",
  height = 15,
  width = 15,
  edge_type = "arc",
  region_legend = TRUE,
  degree_scale_limit = c(1, 10),
  anatomical.colors = NULL,
  correlation_edge_width_limit = c(0.8, 1),
  image_ext = ".png",
  print_plot = TRUE,
  graph_theme = NULL,
  label_size = 5,
  edge_thickness_range = c(1, 5),
  node_size_range = c(1, 8),
  label_offset = 0.15,
  save_plot = TRUE
)
}
\arguments{
\item{e}{experiment object}

\item{network_name}{(str, default = "AD")}

\item{title}{(str, default = NULL) Title of network plot}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel"))}

\item{edge_color}{(str, default = "firebrick") Color of the network edges.}

\item{height}{Height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{edge_type}{(default = "arc") "arc" or "diagonal".
Can also be a hexadecimal color code written as a string.}

\item{region_legend}{(default = TRUE) Boolean determining whether or not to show the region legend categorizing subregions into their largest parent region. Only works well if the Allen ontology is used for the dataset.}

\item{degree_scale_limit}{(vec, default = c(1,10)) Scale limit for degree size}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{print_plot}{(bool, default = TRUE) Whether to print the plot as an output.
the experiment object output folder.}

\item{graph_theme}{(default = NULL) Add a \code{\link[ggraph:theme]{ggraph::theme()}} to the network graph. If NULL, the default is taken.}

\item{label_size}{(default = 5) Default font size for network region labels.}

\item{label_offset}{(default = 0.15) Distance of label from nodes.}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}

\item{anatomical.color}{(vec, NULL) Custom colors for the super region legend. (e.g., c(Isocortex = "#5571a9", OLF = "#64bdc4", HPF = "#d2875b",
CTXsp = "#87a3db", CNU = "#466496", TH = "#7e72af", HY = "#8e7960",  MB = "#d796c8", HB = "#646464")). NULL results in viridis.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Plot the networks stored in an experiment object
}
