% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_normalized_counts}
\alias{plot_normalized_counts}
\title{Plot normalized cell counts}
\usage{
plot_normalized_counts(
  e,
  channels = c("cfos", "eyfp", "colabel"),
  by = c("sex", "group"),
  values = list(c("female", "non"), c("female", "agg"), c("female", "control"),
    c("male", "agg"), c("male", "control")),
  colors = c("white", "lightblue", "black", "red", "green"),
  ontology = "allen",
  title = NULL,
  unit_label = bquote(`Cell counts `("cells/mm"^3)),
  anatomical.order = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU", "TH", "HY", "MB",
    "HB", "CB"),
  height = 7,
  width = 20,
  region_label_angle = 50,
  label_text_size = 8,
  print_plot = TRUE,
  save_plot = TRUE,
  flip_axis = FALSE,
  reverse_colors = FALSE,
  legend.justification = c(0, 0),
  legend.position = "inside",
  legend.position.inside = c(0.05, 0.6),
  legend.direction = "vertical",
  limits = c(0, 1e+05),
  facet_background_color = NULL,
  strip_background_color = "lightblue",
  image_ext = ".pdf"
)
}
\arguments{
\item{e}{experiment object}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel"))}

\item{by}{(str) Attribute names to group by, e.g. c("sex", "group")}

\item{values}{(list) A list with a length the number of groups desired for plotting. Each element of the list is a vector in the order of the
the respective values for the attributes entered for the \code{by} parameter to generate a specific analysis group. Each vector should be unique to generate a
uniquely colored bar.
e.g.values = c("female", "AD").}

\item{colors}{(str, default = c("white", "lightblue")) Hexadecimal codes corresponding to the groups (respectively) to plot. The length of this vector should be the length of the list.}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}

\item{title}{(str, default = NULL) An optional title for the plot}

\item{unit_label}{(str, default = bquote('Cell counts '('cells/mm'^3))) Default unit label for the graphs}

\item{anatomical.order}{(default = c("Isocortex", "OLF", "HPF", "CTXsp", "CNU","TH", "HY", "MB", "HB", "CB")) Default way to group subregions into super regions order}

\item{height}{height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{region_label_angle}{(int, default = 50) Angle of the region label.}

\item{label_text_size}{(int, default = 8) Size of the text element for the region labels.}

\item{print_plot}{(bool, default = TRUE) Whether to display the plot (in addition to saving the plot)}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of
the experiment object output folder.}

\item{flip_axis}{plot cell counts on x-axis rather than y-axis.}

\item{reverse_colors}{(bool, default = FALSE) Whether to reverse the color order. This may depend on the order in which you entered the \code{colors} parameter}

\item{legend.justification}{(default = c(0, 0))}

\item{legend.position}{(default = "inside")}

\item{legend.position.inside}{(default = c(0.05, 0.6))}

\item{legend.direction}{(c("vertical", "horizontal"))}

\item{limits}{(c(0,100000)) Range of the normalized cell counts.}

\item{facet_background_color}{(default = NULL) Set to a hexadecimal string, e.g."#FFFFFF", when you want to shade the background of the graph. Defaults to no background when NULL.}

\item{strip_background_color}{(default = "lightblue") Color of background srip delineating major anatomical parent regions. Set to color name or to hexadecimal code.}

\item{image_ext}{(default = ".png") image extension to the plot as.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Plot the cell counts normalized by volume for a given channel
}
