% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_normalized_counts}
\alias{plot_normalized_counts}
\title{Plot normalized cell counts}
\usage{
plot_normalized_counts(
  e,
  channels = c("cfos", "eyfp", "colabel"),
  groups = c("Context", "Shock"),
  colors = c("white", "lightblue"),
  regions_to_remove = c("CTX", "grey", "MB", "TH", "HY"),
  title = NULL,
  unit_label = bquote(`Cell counts `("cells/mm"^3)),
  height = 7,
  width = 20,
  region_label_angle = 50,
  label_text_size = 8,
  print_plot = TRUE,
  save_plot = TRUE,
  flip_axis = FALSE,
  facet_background_color = NULL,
  image_ext = ".pdf"
)
}
\arguments{
\item{e}{experiment object}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel"))}

\item{groups}{(str, default = c("Context", "Shock")) The groups to be plotted. The order of this vector will dictate the ordering of grouped bars in the resulting plot.}

\item{colors}{(str, default = c("white", "lightblue")) Hexadecimal codes corresponding to the groups (respectively) to plot.}

\item{regions_to_remove}{(str, default = c("CTX", "grey", "MB", "TH", "HY")) A vector of strings corresponding to regions to remove from the plot. The default regions are parent regions from which their subregions have already been plotted. Set to NULL,
FALSE, NULL or NA  to avoid removing regions.}

\item{title}{(str, default = NULL) An optional title for the plot}

\item{unit_label}{(str, default = bquote('Cell counts '('cells/mm'^3))) Default unit label for the graphs}

\item{height}{height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{region_label_angle}{(int, default = 50) Angle of the region label.}

\item{label_text_size}{(int, default = 8) Size of the text element for the region labels.}

\item{print_plot}{(bool, default = TRUE) Whether to display the plot (in addition to saving the plot)}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of
the experiment object output folder.}

\item{flip_axis}{plot cell counts on x-axis rather than y-axis.}

\item{facet_background_color}{(default = NULL) Set to a hexadecimal string, e.g."#FFFFFF", when you want to shade the background of the graph. Defaults to no background when NULL.}

\item{image_ext}{(default = ".png") image extension to the plot as.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Plot the cell counts normalized by volume for a given channel
}
