% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_percent_colabel}
\alias{plot_percent_colabel}
\title{plot_colabel_percent}
\usage{
plot_percent_colabel(
  e,
  by = NULL,
  channel = "cfos",
  rois = c("AAA", "dDG", "HY"),
  color_mapping = "sex",
  colors = c("#952899", "#358a9c"),
  pattern_mapping = "group",
  patterns = c("gray100", "hs_fdiagonal", "hs_horizontal", "gray90", "hs_vertical"),
  error_bar = "sem",
  ylim = c(0, 100),
  plot_individual = TRUE,
  height = 10,
  width = 10,
  print_plot = TRUE,
  save_plot = TRUE,
  image_ext = ".png"
)
}
\arguments{
\item{e}{experiment object}

\item{error_bar}{(str, c("sd", "sem)) options for which type of error bar to display, standard deviation or standard error of the mean.}
}
\value{

}
\description{
plot_colabel_percent
}
