% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{plot_percent_colabel}
\alias{plot_percent_colabel}
\title{This function allows for plotting of colabelled cells over either the "cfos" or "eyfp" channels. And allows for specification
of specific brain regions to plot. Two different mouse attributes can be used as categorical variables to map to either the color or
pattern aesthetics of the bar plot, e.g. sex and experimental group.
The color aesthetic takes precedence over the pattern aesthetic so if you only want to use one mouse attribute, for plotting
set it to the \code{color_mapping} parameter and set the \code{pattern_mapping} parameter to NULL.}
\usage{
plot_percent_colabel(
  e,
  colabel_channel = "colabel",
  channel = "eyfp",
  rois = c("AAA", "dDG", "HY"),
  color_mapping = "sex",
  colors = c("#952899", "#358a9c"),
  pattern_mapping = NULL,
  patterns = c("gray100", "hs_fdiagonal",
    "hs_horizontal\\n                                              ", "gray90",
    "hs_vertical"),
  error_bar = "sem",
  ylim = c(0, 100),
  plot_individual = TRUE,
  height = 8,
  width = 8,
  print_plot = TRUE,
  save_plot = TRUE,
  image_ext = ".png"
)
}
\arguments{
\item{e}{experiment object}

\item{channel}{(str, default = "eyfp") The channel used as denominator in fraction counts.}

\item{rois}{character vector of region acronyms, e.g. c("AAA", "DG)}

\item{color_mapping}{(str, default = "sex") The name of the categorical variable (e.g., "sex", "age", etc.) to map to the color aesthetic of the bar plot.}

\item{colors}{(str) character vector of the color values desired for the groups.}

\item{pattern_mapping}{(str, default = "sex") The name of the categorical variable (e.g., "sex", "age", etc.) to map to the pattern aesthetic of the bar plot.}

\item{patterns}{(str, default = c("gray100", 'hs_fdiagonal', "hs_horizontal", "gray90", "hs_vertical"), Pattern types to define subgroups.}

\item{error_bar}{(str, c("sd", "sem)) options for which type of error bar to display, standard deviation or standard error of the mean.}
}
\value{
p Plot handle to the figure
}
\description{
This function allows for plotting of colabelled cells over either the "cfos" or "eyfp" channels. And allows for specification
of specific brain regions to plot. Two different mouse attributes can be used as categorical variables to map to either the color or
pattern aesthetics of the bar plot, e.g. sex and experimental group.
The color aesthetic takes precedence over the pattern aesthetic so if you only want to use one mouse attribute, for plotting
set it to the \code{color_mapping} parameter and set the \code{pattern_mapping} parameter to NULL.
}
\examples{
plot_percentage_colabel
}
