% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.experiment}
\alias{print.experiment}
\title{Print attributes of experiment object}
\usage{
\method{print}{experiment}(x, ...)
}
\arguments{
\item{x}{experiment object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print attributes of experiment object
}
\examples{
e <- experiment()
print(e)
}
