% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{print.mouse}
\alias{print.mouse}
\title{Print attributes of mouse object}
\usage{
\method{print}{mouse}(m)
}
\arguments{
\item{m}{mouse object}
}
\description{
Print attributes of mouse object
}
