% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_mouse}
\alias{save_mouse}
\title{Save mouse data}
\usage{
save_mouse(..., timestamp = FALSE)
}
\arguments{
\item{...}{parameter to pass mouse object}

\item{timestamp}{(bool) save the object with a date tag}
}
\description{
Saves mouse object into it's attribute output path as an RDATA file
save_mouse(m)
}
\examples{
\dontrun{
m <- save_mouse(m, timestamp = TRUE)
}
}
