% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping_functions.R
\name{save_mouse}
\alias{save_mouse}
\title{Save mouse data}
\usage{
save_mouse(m)
}
\arguments{
\item{m}{mouse object}
}
\description{
Saves mouse data into it's attribute output path as an RDATA file
}
