% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{segmentation.object}
\alias{segmentation.object}
\title{segmentation object compatible with wholebrain package functions}
\format{
A
\describe{
\item{filter}{list storing parameter use to segment and get brain contours}
\item{soma}{list storing cell count data}
}
}
\usage{
segmentation.object
}
\description{
segmentation object compatible with wholebrain package functions
}
\keyword{datasets}
