% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simplify_vec_by_keywords}
\alias{simplify_vec_by_keywords}
\title{Simplify vector of acronyms by keywords.}
\usage{
simplify_vec_by_keywords(
  vec,
  keywords = c("layer", "part", "stratum", "division", "leaflet", "Subgeniculate",
    "island", "Islands", "Fields of Forel", "Cajal", "Darkschewitsch", "Precommissural"),
  ontology = "allen",
  dont_fold = c("Dorsal part of the lateral geniculate complex",
    "Ventral posterolateral nucleus of the thalamus, parvicellular part",
    "Ventral posteromedial nucleus of the thalamus, parvicellular part",
    "Ventral posterolateral nucleus of the thalamus, parvicellular part",
    "Ventral posteromedial nucleus of the thalamus, parvicellular part",
    "Substantia nigra")
)
}
\arguments{
\item{vec}{(vector) Must contain acronyms}

\item{keywords}{(vec, default = c("layer","part","stratum","division")) a list of keywords to simplify based on region name.}

\item{ontology}{(str, default = "allen") Region ontology to use. options = "allen" or "unified"}

\item{dont_fold}{(vec, default = c("Dorsal part of the lateral geniculate complex", "Ventral posterolateral nucleus of the thalamus, parvicellular part", "
Ventral posteromedial nucleus of the thalamus, parvicellular part","Ventral posterolateral nucleus of the thalamus, parvicellular part",
"Ventral posteromedial nucleus of the thalamus, parvicellular part","Substantia nigra")) Regions that are exceptions to being folded into their parent regions.}
}
\value{
df, dataframe as a tibble with included long name and acronyms that are simplified to parents
}
\description{
Simplify vector of acronyms by keywords.
}
\examples{
acronyms <- c("LGd", "GU4", "dCA1so" )
simplify_vec_by_keywords(acronyms)

acronyms <- c("LPBD", "MPBE", "CPre")
simplify_vec_by_keywords(acronyms, keywords = c("dorsal part", "external part", "Caudoputamen-"), ontology = "unified")
}
