% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTR.R
\docType{package}
\name{SMARTR}
\alias{SMARTR}
\title{SMARTR: A mapping, analysis, and visualization package for wholebrain dual-ensemble coronal datasets.}
\description{
The base of this pipeline is an R package in with the title, \code{SMARTR}, a self-referential play on
a previous package developed as an extension to \code{wholebrain} called \code{\link{SMART}}.  This package allows for the
user-friendly pre-processing of segmentation data generated from ImageJ to a be compatible with the \code{wholebrain} package to generate
region-based cell counts that are normalized by volume. It will also provides tools for data analysis based on experimental groupings.
}
\details{
Object descriptions

The data for analysis will be stored in objects that allow for more neat bundling of useful information together.

A \link{slice} object will contain all the data related to registration, segmentation for each channel, and cell counts for a particular image.
It will also contain “metadata” about your experimental images, such as what the experimenter-assigned slice ID is,
which brain atlas AP coordinate matches best with the given image, and what the path to the image used for registration is.
These metadata are stored as the object’s attributes.

A \link{mouse} object is an object that will store multiple slice objects (and therefore all the information in it),
and will eventually store the combined cell data and the region cell counts normalized by volume.
Like a slice, it will also contain “metadata” about your mouse stored as attributes.
An \link{experiment}An experiment object consists of a list of processed mouse objects with raw data from slices omitted, and experimental attributes stored as a list.
It will also contain “metadata” about your experimental personnel and analysis groups stored as attributes.
}
\section{The package currently allows for easy implementation of the following steps}{

\enumerate{
\item Setting up the pipeline by specifying experimentparameters, and save directories.
\item The interactive registration process.
\item Importing raw segmentation data from .txt files generated from ImageJ for multiple channels.
\item Optionally creating a filter for the 'cfos' and 'eyfp' channels to clean segmented counts.
\item Creating a segmentation object that is compatible with \code{wholebrain} functions.
\item Forward warping and mapping the data onto the standardized mouse atlas.
\item Cleaning the mapped data in all the following ways: + Removing cells that map outside the boundaries of the atlas.
\itemize{
\item Omitting regions by a default list of regions to omit.
\item Omitting regions by user specified region acronyms.
\item Removing Layer 1 cells
\item Removing cells from a contralateral hemisphere per slice if the registrations are divided by right and left hemispheres.
}
\item Obtaining cell counts normalized by region volume (per mm^2^) and region areas (per mm^2^).
}
}

