% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{split_hipp_DV}
\alias{split_hipp_DV}
\title{Split the hippocampal dataset to dorsal and ventral regions.}
\usage{
split_hipp_DV(
  m,
  AP_coord = -2.7,
  combine_hemispheres = TRUE,
  merge = TRUE,
  simplify_regions = TRUE,
  simplify_keywords = c("layer", "part", "stratum", "division"),
  rois = c("DG", "CA1", "CA2", "CA3")
)
}
\arguments{
\item{m}{mouse object}

\item{AP_coord}{(double) The AP coordinate which separates dorsal hippocampus from ventral hippocampus. Counts anterior to this are considered dorsal, counts posterior to
this are considered ventral.}

\item{combine_hemispheres}{(bool, default = TRUE) Combine normalized cell counts from both hemispheres}

\item{merge}{(bool, default = TRUE) Recommended setting. Whether to include the division of dHipp and vHipp into the normalized counts dataframe of the mouse or to store as a separate element for isolated analysis.
Merging will replace the hippocampal counts currently in the normalized counts table. Not merging will create a separate data element in the mouse. TODO: Currently not functional.}

\item{simplify_regions}{(bool, default = TRUE ) simplify the normalized region counts based on keywords in the internal function, \code{simplify_keywords}}

\item{simplify_keywords}{(str, default =  c("layer","part","stratum","division")). Keywords to search through region names and simplify to parent structure}

\item{rois}{(str, default = c("DG", "CA1", "CA2", "CA3")) List of the region acronyms to include in the hippocampus}
}
\value{
m mouse object with DV cell counts from the hippocampus included
}
\description{
If \code{\link[=normalize_cell_counts]{normalize_cell_counts()}} has already been run,
the existing hippocampus data will be deleted and either replaced with new dorsal/ventral hippocampus data when merge = TRUE (recommended) or
it will be stored in a separate dataframe if the user wants analyze the hippocampus separately for analysis and the \code{normalized_counts} dataframe will be unchanged.
}
