% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{summarize_null_networks}
\alias{summarize_null_networks}
\title{Summarize the parameters of the rewired null networks generated by \code{\link[=rewire_network]{rewire_network()}}}
\usage{
summarize_null_networks(
  null_nodes_p1,
  null_nodes_p2,
  network_name_1 = p[[1]],
  network_name_2 = p[[2]],
  channel = "cfos"
)
}
\arguments{
\item{null_nodes_p1}{Summary table of rewired network properties of all nodes from \code{\link[=rewire_network]{rewire_network()}} for the first network to compare.}

\item{null_nodes_p2}{Summary table of rewired network properties of all nodes from \code{\link[=rewire_network]{rewire_network()}} for the second network to compare}

\item{network_name_1}{(str) Name of the network}

\item{network_name_2}{(str) Name of the network}

\item{channel}{(str)   channel to process}
}
\value{
a list of length 2. The first element is named \code{global_summary} and contains a table of global summary statistics.
The second element is named \code{node_summary}, and contained per node statistics averaged from multiple null networks.
}
\description{
Summarize the parameters of the rewired null networks generated by \code{\link[=rewire_network]{rewire_network()}}
}
\examples{
\dontrun{
null_nodes_p1 <- rewire_network(anesthesia,
                                channels = "cfos",
                                network_name = "Ctrl",
                                n_rewires = igraph::gsize(anesthesia$networks$Ctrl$cfos)*100,  # no. edge rewires = no. edges in the graph x 100
                                return_graphs = FALSE)
}
}
