% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Plot the results of the permutation histogram used to determine the p-value of the pairwise region comparison}
\usage{
volcano_plot(
  e,
  permutation_comparison = "female_AD_vs_male_AD",
  channels = c("cfos", "eyfp", "colabel"),
  colors = c("#be0000", "#00782e", "#f09b08"),
  save_plot = TRUE,
  title = NULL,
  ylim = c(0, 3),
  height = 8,
  width = 10,
  print_plot = FALSE,
  plt_theme = NULL,
  point_size = 1,
  image_ext = ".png"
)
}
\arguments{
\item{e}{experiment object}

\item{permutation_comparison}{The name of the correlation group comparisons to plot.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) channels to plot.}

\item{colors}{(str, default = c("#be0000", "#00782e", "#f09b08")) Hexadecimal code for the colors corresponding to the
channels attribute of the correlation_list. Color values can also be input compatible with ggplot2 plotting functions.}

\item{save_plot}{(bool, default = TRUE) Save into the figures subdirectory of the
the experiment object output folder.}

\item{title}{Title of the plot.}

\item{ylim}{(vec, default = c(0,y)) Y-axis range (logarithmic).}

\item{height}{height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{print_plot}{(bool, default = FALSE) Print the plot as graphics windows.}

\item{plt_theme}{(default = NULL) Add a \code{\link[ggplot2:theme]{ggplot2::theme()}} to the plot. If NULL, the default is taken..}

\item{point_size}{(default = 1) Size of the plotted points.}

\item{image_ext}{(default = ".png") image extension to save the plot as.}
}
\value{
p_list A list the same length as the number of channels, with each element containing a plot handle for that channel.
}
\description{
Create a Volcano plot.
}
\details{
Plot the correlation difference between two comparison groups into a volcano plot. The function
\code{\link[=correlation_diff_permutation]{correlation_diff_permutation()}} must be run first in order to generate results to plot.
}
\examples{
volcano_plot(e, permutation_comparison = "female_AD_vs_male_AD", channels = c("cfos", "eyfp", "colabel"),
colors =  c("#be0000", "#00782e", "#f09b08"), save_plot = TRUE, title = NULL, ylim = c(0, 3), height = 8,
width = 10, print_plot = FALSE, image_ext = ".png")
}
