% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_plotting_functions.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Create a Volcano plot.}
\usage{
volcano_plot(
  e,
  permutation_comparison = "female_AD_vs_male_AD",
  channels = c("cfos", "eyfp", "colabel"),
  colors = c("#be0000", "#00782e", "#f09b08"),
  alpha = 0.05,
  save_plot = TRUE,
  title = NULL,
  height = 8,
  width = 10,
  print_plot = TRUE,
  image_ext = ".png"
)
}
\arguments{
\item{permutation_comparison}{The name of the correlation group comparisons to plot.}

\item{channels}{(str, default = c("cfos", "eyfp", "colabel")) channels to plot}

\item{colors}{(str, default = c("#be0000", "#00782e", "#f09b08")) Hexadecimal codes corresponding to the channels (respectively) to plot.}

\item{alpha}{(float, default = 0.05) The alpha value used to plot horizontal line. Point above the line are considered significant.}

\item{save_plot}{(bool, default = TRUE) Save the correlation heatmap plots into the figures subdirectory of the}

\item{title}{Title of the plot.}

\item{height}{height of the plot in inches.}

\item{width}{width of the plot in inches.}

\item{image_ext}{(default = ".png") image extension to the plot as.}

\item{store}{(bool, default = FALSE) Store the plot object into your experiment.}
}
\value{
e experiment object
}
\description{
Plot the correlation difference between two comparison groups into a volcano plot. The function
\code{\link[=correlation_diff_permutation]{correlation_diff_permutation()}} must be run first in order to generate results to plot.
}
